feat: drop rooms.h circural dependency on helapordo.h
