+++ src/game_curses.c
@@ -1632,222 +1632,6 @@ void boss_print_in_panel(WINDOW *win, int starty, int startx, int width,
 }
 
 /**
- * Takes a WINDOW pointer to print notifications to, a fighterStatus value and a string of who's the entity to print the respective status message.
- * @see fighterStatus
- * @param notify_win The pointer to the window to use display_notification() on.
- * @param status The fighterStatus at hand.
- * @param subject A string with name of entity owning the fighterStatus.
- */
-void printStatusText(WINDOW *notify_win, fighterStatus status, char *subject)
-{
-    char msg[500];
-    switch (status) {
-    case Normal: {
-        return;
-    };
-    break;
-    case Poison:
-    case Burned: {
-        sprintf(msg, "%s is hurt by its %s.", subject,
-                stringFromStatus(status));
-        display_notification(notify_win, msg, 500);
-    }
-    break;
-    case Weak:
-    case Strong: {
-        sprintf(msg, "%s is feeling %s.", subject,
-                stringFromStatus(status));
-        display_notification(notify_win, msg, 500);
-    }
-    break;
-    case Frozen: {
-        sprintf(msg, "%s is frozen cold.", subject);
-        display_notification(notify_win, msg, 500);
-    }
-    break;
-    }
-}
-
-/**
- * Takes a WINDOW pointer to print notifications to, a Fighter pointer value and applies the effect pertaining to its status value.
- * @see Fighter
- * @see fighterStatus
- * @see printStatusText()
- * @param notify_win The WINDOW pointer to call display_notification() on.
- * @param f The Fighter pointer at hand.
- */
-void applyStatus(WINDOW *notify_win, Fighter *f)
-{
-
-    switch (f->status) {
-    case Normal: {
-        break;
-    }
-    break;
-    case Poison: {
-
-        //Account for penicillin perk
-        //ATM multiples don't stack
-        int penicillin = f->perks[PENICILLIN]->innerValue;
-        if (penicillin > 0) {
-            return;
-        }
-
-        if (f->hp >= 4) {
-            f->hp -= 3;
-        } else {
-            f->hp = 1;	//Will this be a problem?
-        }
-        printStatusText(notify_win, Poison, f->name);
-    }
-    break;
-    case Burned: {
-        printStatusText(notify_win, Burned, f->name);
-    }
-    break;
-    case Frozen: {
-        printStatusText(notify_win, Frozen, f->name);
-    }
-    break;
-    case Weak:
-
-        break;
-    case Strong:
-
-        break;
-    }
-}
-
-/**
- * Takes a WINDOW pointer to print notifications to, a Enemy pointer value and applies the effect pertaining to its status value.
- * @see Enemy
- * @see fighterStatus
- * @see printStatusText()
- * @see stringFromEClass()
- * @param notify_win The window pointer to call display_notification() on.
- * @param e The Enemy pointer at hand.
- * @see display_notification()
- */
-void applyEStatus(WINDOW *notify_win, Enemy *e)
-{
-
-    wattron(notify_win, COLOR_PAIR(S4C_BRIGHT_GREEN));
-
-    switch (e->status) {
-    case Normal: {
-        break;
-    }
-    break;
-    case Poison: {
-        if (e->hp >= 4) {
-            e->hp -= 3;
-        } else {
-            e->hp = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-        printStatusText(notify_win, Poison, stringFromEClass(e->class));
-    }
-    break;
-    case Burned: {
-        if (e->hp >= 5) {
-            e->hp -= 4;
-        } else {
-            e->hp = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-
-        if (e->atk >= 3) {
-            e->atk -= 3;
-        } else {
-            e->atk = 1;
-        }
-        printStatusText(notify_win, Burned, stringFromEClass(e->class));
-    }
-    break;
-    case Frozen: {
-        if (e->vel >= 3) {
-            e->vel -= 1;
-        } else {
-            e->vel = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-        printStatusText(notify_win, Frozen, stringFromEClass(e->class));
-    }
-
-    break;
-    case Weak:
-
-        break;
-    case Strong:
-
-        break;
-    }
-
-    wattroff(notify_win, COLOR_PAIR(S4C_BRIGHT_GREEN));
-}
-
-/**
- * Takes a WINDOW pointer to print notifications to, a Boss pointer value and applies the effect pertaining to its status value.
- * @see Boss
- * @see fighterStatus
- * @see printStatusText()
- * @see stringFromBossClass()
- * @param notify_win The window pointer to call disaply_notification() on.
- * @param b The Boss pointer at hand.
- */
-void applyBStatus(WINDOW *notify_win, Boss *b)
-{
-
-    wattron(notify_win, COLOR_PAIR(S4C_BRIGHT_GREEN));
-
-    switch (b->status) {
-    case Normal: {
-        break;
-    }
-    break;
-    case Poison: {
-        if (b->hp >= 4) {
-            b->hp -= 3;
-        } else {
-            b->hp = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-        printStatusText(notify_win, Poison, stringFromBossClass(b->class));
-    }
-    break;
-    case Burned: {
-        if (b->hp >= 5) {
-            b->hp -= 4;
-        } else {
-            b->hp = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-
-        if (b->atk >= 3) {
-            b->atk -= 3;
-        } else {
-            b->atk = 1;
-        }
-        printStatusText(notify_win, Burned, stringFromBossClass(b->class));
-    }
-    break;
-    case Frozen: {
-        if (b->vel >= 3) {
-            b->vel -= 1;
-        } else {
-            b->vel = 1;	//Will this be a problem for kills in the enemy loop?
-        }
-        printStatusText(notify_win, Frozen, stringFromBossClass(b->class));
-    }
-
-    break;
-    case Weak:
-
-        break;
-    case Strong:
-
-        break;
-    }
-
-    wattroff(notify_win, COLOR_PAIR(S4C_BRIGHT_GREEN));
-}
-
-/**
  * Takes a Enemy pointer and prepares its sprite field by copying it line by line from enemies_sprites, defined in sprites.h header.
  * @see Enemy
  * @see initEnemyStats
