+++ src/helapordo.c
@@ -17,736 +17,6 @@
 */
 #include "helapordo.h"
 
-/**
- * Takes a turnOption_OP and a pointer to a turnOP_args struct.
- * Logs a warning for any NULL field in the struct.
- * Performs the defined turn operation, before returning an OP_res.
- * @param op The kind of operation to do.
- * @param args Struct containing needed args for current operation. Can have some fields uninitialised, if not relevant to requested turnOP.
- * @param kls The Koliseo used for allocations.
- * @param t_kls The Koliseo_Temp used for temporary allocations.
- * @return An OP_res representing result of turn option operation.
- * @see turnOP_args
- * @see turnOption_OP
- * @see OP_res
- */
-OP_res turnOP(turnOption_OP op, turnOP_args *args, Koliseo *kls,
-              Koliseo_Temp *t_kls)
-{
-
-    OP_res res = INVALID_OP;
-
-    Gamestate *gmst = args->gmst;
-    if (gmst == NULL)
-        log_tag("debug_log.txt", "[WARN]", "turnOP_args->(gmst) was NULL");
-    Fighter *actor = args->actor;
-    if (actor == NULL)
-        log_tag("debug_log.txt", "[WARN]", "turnOP_args->(actor) was NULL");
-    Path *path = args->path;
-    if (path == NULL)
-        log_tag("debug_log.txt", "[WARN]", "turnOP_args->(path) was NULL");
-    Room *room = args->room;
-    if (room == NULL)
-        log_tag("debug_log.txt", "[WARN]", "turnOP_args->(room) was NULL");
-    loadInfo *load_info = args->load_info;
-    if (load_info == NULL)
-        log_tag("debug_log.txt", "[WARN]", "turnOP_args->(load_info) was NULL");
-    Enemy *enemy = args->enemy;
-    if (enemy == NULL)
-        log_tag("debug_log.txt", "[WARN]", "turnOP_args->(enemy) was NULL");
-    Boss *boss = args->boss;
-    if (boss == NULL)
-        log_tag("debug_log.txt", "[WARN]", "turnOP_args->(boss) was NULL");
-    FILE *save_file = args->save_file;
-    if (save_file == NULL)
-        log_tag("debug_log.txt", "[WARN]", "turnOP_args->(save_file) was NULL");
-#ifdef HELAPORDO_CURSES_BUILD
-    WINDOW *notify_win = args->notify_win;
-#else
-#ifndef HELAPORDO_RAYLIB_BUILD
-#error "HELAPORDO_CURSES_BUILD and HELAPORDO_RAYLIB_BUILD are both undefined.\n"
-#else
-    Rectangle *notify_win = args->notify_win;
-#endif // HELAPORDO_RAYLIB_BUILD
-#endif // HELAPORDO_CURSES_BUILD
-    if (notify_win == NULL)
-        log_tag("debug_log.txt", "[WARN]",
-                "turnOP_args->(notify_win) was NULL");
-    foeTurnOption_OP foe_op = args->foe_op;
-    if (foe_op == FOE_OP_INVALID) {
-        log_tag("debug_log.txt", "[WARN]",
-                "turnOP_args->(foe_op) was FOE_OP_INVALID");
-    } else if (foe_op < 0 || foe_op > FOETURNOP_MAX) {
-        log_tag("debug_log.txt", "[ERROR]",
-                "turnOP_args->(foe_op) was invalid: [%i]", foe_op);
-        kls_free(default_kls);
-        kls_free(temporary_kls);
-        exit(EXIT_SUCCESS);
-    }
-    skillType skill = args->picked_skill;
-    if (skill < 0 || skill >= SKILLSTOTAL) {
-        log_tag("debug_log.txt", "[WARN]",
-                "turnOP_args->(picked_skill) was invalid: [%i]", skill);
-    }
-
-    int isBoss = -1;
-    int room_index = -1;
-    int enemy_index = -1;
-
-    //Log operation code
-    log_OP(op);
-    switch (op) {
-    case OP_INVALID: {
-        //No_OP, but unexpected
-        log_tag("debug_log.txt", "[ERROR]",
-                "Done turnOP() on an OP_INVALID opcode.");
-    }
-    break;
-    case OP_NEW_GAME: {
-        res = NO_OP;
-    }
-    break;
-    case OP_LOAD_GAME: {
-        if (load_info == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "turnOP_args->(load_info) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        load_info->is_new_game = 0;
-        res = OP_OK;
-    }
-    break;
-    case OP_LOAD_ENEMYROOM: {
-        if (load_info == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "turnOP_args->(load_info) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        load_info = args->load_info;
-        if (load_info == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "load_info was NULL after load_info = args->load_info:  in turnOP(OP_LOAD_ENEMYROOM");
-            return res;
-        }
-        save_file = args->save_file;
-
-        if (save_file == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "turnOP_args->(save_file) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        int *ptr_to_loaded_enemy_index = &(load_info->enemy_index);
-        log_tag("debug_log.txt", "[TURNOP]",
-                "*(ptr_to_loaded_enemy_index) == [%i]",
-                *ptr_to_loaded_enemy_index);
-        int *ptr_to_loaded_roomtotalenemies =
-            (load_info->ptr_to_roomtotalenemies);
-        log_tag("debug_log.txt", "[TURNOP]",
-                "*(ptr_to_loaded_roomtotalenemies) == [%i]",
-                *ptr_to_loaded_roomtotalenemies);
-        int *ptr_to_loaded_roomindex = (load_info->ptr_to_roomindex);
-        log_tag("debug_log.txt", "[TURNOP]",
-                "*(ptr_to_loaded_roomindex) == [%i]",
-                *ptr_to_loaded_roomindex);
-        int *tot_foes = &(load_info->total_foes);
-        log_tag("debug_log.txt", "[TURNOP]", "*(tot_foes) == [%i]",
-                *tot_foes);
-        int *ptr_to_done_loading = &(load_info->done_loading);
-        log_tag("debug_log.txt", "[TURNOP]", "*(done_loading) == [%i]",
-                *ptr_to_done_loading);
-        res =
-            handleLoadgame_Enemies(save_file, actor, path,
-                                   load_info->loaded_enemy,
-                                   ptr_to_loaded_enemy_index,
-                                   ptr_to_loaded_roomtotalenemies,
-                                   ptr_to_loaded_roomindex, tot_foes,
-                                   ptr_to_done_loading, kls);
-        //Log end of operation
-        log_tag("debug_log.txt", "[TURNOP]",
-                "Done operation: [%s] res: [%s (%i)]", stringFromTurnOP(op),
-                stringFrom_OP_res(res), res);
-
-        log_tag(OPS_LOGFILE, "[RES]", "res: [%s (%i)]",
-                stringFrom_OP_res(res), res);
-
-        return res;
-    }
-    break;
-    case OP_LOAD_HOMEROOM: {
-        if (load_info == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "turnOP_args->(load_info) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        load_info = args->load_info;
-        if (load_info == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "load_info was NULL after load_info = args->load_info:  in turnOP(OP_LOAD_HOMEROOM");
-            return res;
-        }
-        save_file = args->save_file;
-
-        if (save_file == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "turnOP_args->(save_file) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        if (load_info->ptr_to_roomindex == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "load_info->ptr_to_roomindex was NULL in turnOP(OP_LOAD_HOMEROOM");
-            return res;
-        }
-        int *ptr_to_loaded_roomindex = (load_info->ptr_to_roomindex);
-        if (ptr_to_loaded_roomindex == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "ptr_to_loaded_roomindex was NULL in turnOP(OP_LOAD_HOMEROOM");
-            return res;
-        }
-        log_tag("debug_log.txt", "[TURNOP]",
-                "*(ptr_to_loaded_roomindex) == [%i]",
-                *ptr_to_loaded_roomindex);
-
-        int *ptr_to_done_loading = &(load_info->done_loading);
-        log_tag("debug_log.txt", "[TURNOP]", "*(done_loading) == [%i]",
-                *ptr_to_done_loading);
-        res =
-            handleLoadgame_Home(save_file, actor, path,
-                                ptr_to_loaded_roomindex,
-                                ptr_to_done_loading, kls);
-        //Log end of operation
-        log_tag("debug_log.txt", "[TURNOP]",
-                "Done operation: [%s] res: [%s (%i)]", stringFromTurnOP(op),
-                stringFrom_OP_res(res), res);
-
-        log_tag(OPS_LOGFILE, "[RES]", "res: [%s (%i)]",
-                stringFrom_OP_res(res), res);
-
-        return res;
-    }
-    break;
-    case OP_FIGHT: {
-        if (notify_win == NULL) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "Notification WINDOW pointer was null in turnOP(OP_FIGHT)");
-            exit(EXIT_FAILURE);
-        }
-        if (room == NULL) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "Room pointer was null in turnOP(OP_FIGHT)");
-            exit(EXIT_FAILURE);
-        }
-        if (foe_op == FOE_OP_INVALID) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "foe_op was FOE_OP_INVALID in turnOP(OP_FIGHT)");
-            exit(EXIT_FAILURE);
-        } else if (foe_op < 0 || foe_op > FOETURNOP_MAX) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "foe_op was invalid in turnOP(OP_FIGHT): [%i]", foe_op);
-            exit(EXIT_FAILURE);
-        }
-        room_index = room->index;
-        if (room->class == ENEMIES && enemy == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Enemy pointer was null in turnOP(OP_FIGHT) for ENEMIES room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class == BOSS && boss == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Boss pointer was null in turnOP(OP_FIGHT) for BOSS room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class != ENEMIES && room->class != BOSS) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Invalid room class in turnOP(OP_FIGHT): (%s [%i])",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-        }
-        switch (room->class) {
-        case ENEMIES: {
-            enemy_index = enemy->index;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (%i) (OP_FIGHT), isBoss == 0",
-                    enemy->index);
-            isBoss = 0;
-            res =
-                OP_res_from_fightResult(defer_fight_enemy
-                                        (actor, enemy, foe_op,
-                                         notify_win, kls));
-        }
-        break;
-        case BOSS: {
-            enemy_index = 0;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (0) (OP_FIGHT), isBoss == 1");
-            isBoss = 1;
-            res =
-                OP_res_from_fightResult(defer_fight_boss
-                                        (actor, boss, path, foe_op,
-                                         notify_win, kls));
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Invalid room value in turnOP(OP_FIGHT): [%s (%i)]",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-
-        }
-        break;
-        }
-    }
-    break;
-    case OP_SPECIAL: {
-        if (room == NULL) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "Room pointer was null in turnOP(OP_SPECIAL)");
-            exit(EXIT_FAILURE);
-        }
-        room_index = room->index;
-        if (room->class == ENEMIES && enemy == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Enemy pointer was null in turnOP(OP_SPECIAL) for ENEMIES room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class == BOSS && boss == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Boss pointer was null in turnOP(OP_SPECIAL) for BOSS room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class != ENEMIES && room->class != BOSS) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Invalid room class in turnOP(OP_SPECIAL): (%s [%i])",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-        }
-        switch (room->class) {
-        case ENEMIES: {
-            enemy_index = enemy->index;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (%i) (OP_SPECIAL), isBoss == 0",
-                    enemy->index);
-            isBoss = 0;
-        }
-        break;
-        case BOSS: {
-            enemy_index = 0;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (0) (OP_SPECIAL), isBoss == 1");
-            isBoss = 1;
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Invalid room value in turnOP(OP_SPECIAL): [%s (%i)]",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-
-        }
-        }
-        handleSpecials(actor, enemy, boss, path, room_index, enemy_index,
-                       isBoss);
-        res = OP_OK;
-    }
-    break;
-    case OP_CONSUMABLE: {
-        if (room == NULL) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "Room pointer was null in turnOP(OP_CONSUMABLE)");
-            exit(EXIT_FAILURE);
-        }
-        room_index = room->index;
-        if (room->class == ENEMIES && enemy == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Enemy pointer was null in turnOP(OP_CONSUMABLE) for ENEMIES room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class == BOSS && boss == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Boss pointer was null in turnOP(OP_CONSUMABLE) for BOSS room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class != ENEMIES && room->class != BOSS) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Invalid room class in turnOP(OP_CONSUMABLE): (%s [%i])",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-        }
-        switch (room->class) {
-        case ENEMIES: {
-            enemy_index = enemy->index;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (%i) (OP_CONSUMABLE), isBoss == 0",
-                    enemy->index);
-            isBoss = 0;
-        }
-        break;
-        case BOSS: {
-            enemy_index = 0;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (0) (OP_CONSUMABLE), isBoss == 1");
-            isBoss = 1;
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Invalid room value in turnOP(OP_CONSUMABLE): [%s (%i)]",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-
-        }
-        }
-        handleConsumables(actor, enemy, boss, isBoss);
-        res = OP_OK;
-    }
-    break;
-    case OP_SAVE: {
-        if (GAMEMODE == Rogue) {
-            log_tag("debug_log.txt", "[WARN]",
-                    "GAMEMODE was [Rogue] in turnOP(OP_SAVE)");
-            res = NO_OP;
-            break;
-        }
-        if (save_file == NULL) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "save_file pointer was null in turnOP(OP_SAVE)");
-            exit(EXIT_FAILURE);
-        }
-        if (load_info == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "turnOP_args->(load_info) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        if (room == NULL) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "Room pointer was null in turnOP(OP_SAVE)");
-            exit(EXIT_FAILURE);
-        }
-        room_index = room->index;
-        if (room->class == ENEMIES && enemy == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Enemy pointer was null in turnOP(OP_SAVE) for ENEMIES room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class == BOSS && boss == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Boss pointer was null in turnOP(OP_SAVE) for BOSS room.");
-            exit(EXIT_FAILURE);
-        }
-
-        if ((room->class != ENEMIES) && (room->class != HOME)) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Invalid room class in turnOP(OP_SAVE): (%s [%i])",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-        }
-
-        switch (room->class) {
-        case ENEMIES: {
-            enemy_index = enemy->index;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (%i) (OP_SAVE), isBoss == 0",
-                    enemy->index);
-            isBoss = 0;
-            load_info->save_type = ENEMIES_SAVE;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting save_type to ENEMIES_SAVE. [%s]",
-                    stringFrom_saveType(load_info->save_type));
-        }
-        break;
-        case BOSS: {
-            enemy_index = 0;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (0) (OP_SAVE), isBoss == 1");
-            isBoss = 1;
-        }
-        break;
-        case HOME: {
-            enemy_index = -1;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (-1) (OP_SAVE), isBoss == -1");
-            isBoss = -1;
-            load_info->save_type = HOME_SAVE;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting save_type to HOME_SAVE. [%s]",
-                    stringFrom_saveType(load_info->save_type));
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Invalid room value in turnOP(OP_SAVE): [%s (%i)]",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-
-        }
-        }
-        switch (load_info->save_type) {
-        case ENEMIES_SAVE: {
-            res =
-                handleSave_Enemies(save_file, actor, path, enemy,
-                                   enemy_index, room->enemyTotal,
-                                   room_index);
-        }
-        break;
-        case HOME_SAVE: {
-            res = handleSave_Home(save_file, actor, path, room_index);
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Invalid save_type value in turnOP(OP_SAVE): (%i)",
-                    (int)load_info->save_type);
-            exit(EXIT_FAILURE);
-        }
-        break;
-        }
-    }
-    break;
-    case OP_DEBUG: {
-        if (gmst == NULL) {
-            log_tag("debug_log.txt", "[WARN]",
-                    "Gamestate pointer was null in turnOP(OP_DEBUG)");
-            exit(EXIT_FAILURE);
-        }
-        if (room == NULL) {
-            log_tag("debug_log.txt", "[WARN]",
-                    "Room pointer was null in turnOP(OP_DEBUG)");
-            exit(EXIT_FAILURE);
-        }
-        room_index = room->index;
-        if (room->class == ENEMIES && enemy == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Enemy pointer was null in turnOP(OP_DEBUG) for ENEMIES room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class == BOSS && boss == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Boss pointer was null in turnOP(OP_DEBUG) for BOSS room.");
-            exit(EXIT_FAILURE);
-        }
-
-        switch (room->class) {
-        case HOME: {
-            enemy_index = -1;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (%i) (OP_DEBUG), isBoss == -1",
-                    enemy_index);
-            isBoss = -1;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Doing endwin() before debug_generic()");
-            endwin();
-            debug_generic(gmst, actor, path, room_index, kls, t_kls);
-            res = OP_OK;
-        }
-        break;
-        case ENEMIES: {
-            enemy_index = enemy->index;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (%i) (OP_DEBUG), isBoss == 0",
-                    enemy->index);
-            isBoss = 0;
-            debug_enemies_room(gmst, room, actor, enemy, path,
-                               room_index, enemy_index, kls, t_kls);
-            res = OP_OK;
-        }
-        break;
-        case BOSS: {
-            enemy_index = 0;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (0) (OP_DEBUG), isBoss == 1");
-            isBoss = 1;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Doing endwin() before debug_generic()");
-            endwin();
-            debug_generic(gmst, actor, path, room_index, kls, t_kls);
-            res = OP_OK;
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Invalid room value in turnOP(OP_DEBUG): [%s (%i)]",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-        }
-        }
-    }
-    break;
-    case OP_EQUIPS: {
-        if (actor == NULL || path == NULL) {
-            if (actor == NULL)
-                log_tag("debug_log.txt", "[ERROR]",
-                        "turnOP_args->(actor) was NULL");
-            if (path == NULL)
-                log_tag("debug_log.txt", "[ERROR]",
-                        "turnOP_args->(path) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        handleEquips(actor, path);
-        res = OP_OK;
-    };
-    break;
-    case OP_PERKS: {
-        if (actor == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "turnOP_args->(actor) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        printActivePerks(actor);
-        res = OP_OK;
-    }
-    break;
-    case OP_STATS: {
-        if (actor == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "turnOP_args->(actor) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        handleStats(actor);
-        res = OP_OK;
-    }
-    break;
-    case OP_ARTIFACTS: {
-        if (actor == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "turnOP_args->(actor) was NULL");
-            //free(args);
-            log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-            return res;
-        }
-        handleArtifacts(actor);
-        res = OP_OK;
-    }
-    break;
-    case OP_EXPLORE: {
-        log_tag("debug_log.txt", "[DEBUG]", "Going to explore!");
-        //The free of args should be done after the OP; in handleRoom_Home()
-        res = NO_OP;
-    }
-    break;
-    case OP_QUIT: {
-        quit(actor, room, load_info, t_kls);
-        //FIXME
-        //We can't free the turnOP_args, can we?
-        //free(args);
-        log_tag("debug_log.txt", "[FREE]", "Freed turnOP_args");
-    }
-    break;
-    case OP_SKILL: {
-
-        log_tag("debug_log.txt", "[DEBUG]", "OP_SKILL");
-
-        res = NO_OP;
-
-        if (notify_win == NULL) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "Notification WINDOW pointer was null in turnOP(OP_SKILL)");
-            exit(EXIT_FAILURE);
-        }
-        if (room == NULL) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "Room pointer was null in turnOP(OP_SKILL)");
-            exit(EXIT_FAILURE);
-        }
-        if (foe_op == FOE_OP_INVALID) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "foe_op was FOE_OP_INVALID in turnOP(OP_SKILL)");
-            exit(EXIT_FAILURE);
-        } else if (foe_op < 0 || foe_op > FOETURNOP_MAX) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "foe_op was invalid in turnOP(OP_SKILL): [%i]", foe_op);
-            exit(EXIT_FAILURE);
-        }
-        if (skill < 0 || skill > SKILLSTOTAL) {
-            log_tag("debug_log.txt", "[CRITICAL]",
-                    "skill was invalid in turnOP(OP_SKILL): [%i]", skill);
-            exit(EXIT_FAILURE);
-        }
-        room_index = room->index;
-        if (room->class == ENEMIES && enemy == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Enemy pointer was null in turnOP(OP_SKILL) for ENEMIES room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class == BOSS && boss == NULL) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Boss pointer was null in turnOP(OP_SKILL) for BOSS room.");
-            exit(EXIT_FAILURE);
-        } else if (room->class != ENEMIES && room->class != BOSS) {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Invalid room class in turnOP(OP_SKILL): (%s [%i])",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-        }
-        switch (room->class) {
-        case ENEMIES: {
-            enemy_index = enemy->index;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (%i) (OP_SKILL), isBoss == 0",
-                    enemy->index);
-            isBoss = 0;
-            //TODO
-            //Implement the missing function to wrap skill usage and foe op
-            res = OP_res_from_fightResult(defer_skill_enemy(actor, enemy, skill, foe_op, notify_win, kls));
-        }
-        break;
-        case BOSS: {
-            enemy_index = 0;
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Setting enemy_index to (0) (OP_SKILL), isBoss == 1");
-            isBoss = 1;
-            //TODO
-            //Implement the missing function to wrap skill usage and foe op
-            res = OP_res_from_fightResult(defer_skill_boss(actor, boss, skill, path, foe_op, notify_win, kls));
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[TURNOP]",
-                    "Invalid room value in turnOP(OP_FIGHT): [%s (%i)]",
-                    stringFromRoom(room->class), room->class);
-            exit(EXIT_FAILURE);
-
-        }
-        break;
-        }
-    }
-    break;
-    default: {
-        log_tag(OPS_LOGFILE, "[ERROR]", "Invalid OP in turnOP()");
-        log_tag("debug_log.txt", "[ERROR]", "Invalid OP in turnOP()");
-    }
-    break;
-    }
-
-    //Log end of operation
-    log_tag("debug_log.txt", "[TURNOP]", "Done operation: [%s] res: [%s (%i)]",
-            stringFromTurnOP(op), stringFrom_OP_res(res), res);
-
-    log_tag(OPS_LOGFILE, "[RES]", "res: [%s (%i)]", stringFrom_OP_res(res),
-            res);
-
-    return res;
-}
-
-fightResult do_Skill(Fighter * player, Enemy * e, skillType picked_skill, WINDOW * notify_win, Koliseo * kls)
-{
-
-    return FIGHTRES_NO_DMG;
-}
-
-fightResult do_Skill_boss(Fighter * player, Boss * b, skillType picked_skill, Path * path, WINDOW * notify_win, Koliseo * kls)
-{
-
-    return FIGHTRES_NO_DMG;
-}
-
 //TODO Drop dead code
 /*
  * Takes an integer, a callback_void_t pointer function and a Fighter holding the array for the callback registration.
@@ -779,146 +49,6 @@ void removeEquipPerks(Equip *e, Fighter *f)
 
         Perk *fighterPerk = f->perks[p->class];
         fighterPerk->innerValue -= 1;
-    }
-}
-
-/**
- * Takes a Figher pointer and prints all of its active perks formatted.
- * @see Fighter
- * @see Perk
- * @see Equip
- * @param f The Fighter pointer with perks to print.
- */
-void printActivePerks(Fighter *f)
-{
-
-    WINDOW *win;
-
-    /* Initialize curses */
-    //initscr();
-    clear();
-    refresh();
-    start_color();
-    cbreak();
-    noecho();
-    keypad(stdscr, TRUE);
-
-    /* Create the windows for player stats and lifetime counters */
-    win = newwin(22, 65, 1, 2);
-    keypad(win, TRUE);
-
-    /* Print a border around the windows and print a title */
-    box(win, 0, 0);
-    print_label(win, 1, 0, 35, "Perks", COLOR_PAIR(S4C_BRIGHT_YELLOW));
-    mvwaddch(win, 2, 0, ACS_LTEE);
-    mvwhline(win, 2, 1, ACS_HLINE, 63);
-    mvwaddch(win, 2, 64, ACS_RTEE);
-
-    wrefresh(win);
-
-    //attron(COLOR_PAIR(3));
-    mvprintw(23, 0, "(Press q or Enter to Exit)");
-    //attroff(COLOR_PAIR(3));
-
-    int y = 4;
-    int x = 2;
-    //int startx = 2;
-    //int width = 65;
-
-    int empty = 1;
-
-    int count = PERKSMAX + 1;
-
-    for (int i = 0; i < count; i++) {
-        Perk *p = f->perks[i];
-        if (p->innerValue > 0) {
-            empty = 0;
-
-            wattron(win, COLOR_PAIR(S4C_CYAN));
-            mvwprintw(win, y, x, " x%i %s ", p->innerValue,
-                      nameStringFromPerk(p->class));
-            wattroff(win, COLOR_PAIR(S4C_CYAN));
-            char s[250];
-            sprintf(s, " x%i %s", p->innerValue, nameStringFromPerk(p->class));
-            int l = strlen(s);
-            wattron(win, COLOR_PAIR(S4C_BRIGHT_YELLOW));
-            mvwprintw(win, y, x + l + 2, "\"%s\"",
-                      descStringFromPerk(p->class));
-            wattroff(win, COLOR_PAIR(S4C_BRIGHT_YELLOW));
-            y++;
-        }
-    };
-
-    if (empty) {		//No perks are active
-        wattron(win, COLOR_PAIR(S4C_BRIGHT_YELLOW));
-        mvwprintw(win, y, x, "You don't have any special power yet.");
-        wattroff(win, COLOR_PAIR(S4C_BRIGHT_YELLOW));
-    }
-
-    refresh();
-
-    int picked = 0;
-    int c = -1;
-    wrefresh(win);
-
-    while (!picked && (c = wgetch(win)) != 'q') {
-        switch (c) {
-        case 10: {		/*Enter */
-            picked = 1;
-
-        };
-        break;
-        }
-    }
-    delwin(win);
-    endwin();
-}
-
-//Counter functions
-
-
-//TODO update this to actually handle different types of counters and reliably print them, maybe to a ncurses window
-/**
- * Takes a Turncounter array.
- * For every Turncounter in the array, the values of count, innerVal, type and all the function pointers fields are printed.
- * @see Turncounter
- * @param counters The Turncounter array to be printed.
- */
-void printCounters(Turncounter *counters[])
-{
-    yellow();
-    printf
-    ("%-10.10s\t%-10.10s\t%-3.3s\t%-3.3s\t%-11.11s\t%-11.11s\t%-11.11s\t%-11.11s\n",
-     "Count", "Desc", "Val", "Typ", "*(eff())", "*(eff_e())", "*(boost())",
-     "*(boost_e())");
-    for (int i = 0; i < (COUNTERSMAX + 1); i++) {
-        Turncounter *c = counters[i];
-        lightBlue();
-        printf("%-10.10i\t%-10.10s\t", c->count, c->desc);
-        strongWhite();
-        printf("(%-3.3i)\t(%-3.3i)\t", c->innerValue, c->type);
-        purple();
-        //printf("[%-11.11i]\t[%-11.11i]\t",*(c->effect_fun),*(c->effect_e_fun));
-        cyan();
-        //printf("[%-11.11i]\t[%-11.11i]\n",*(c->boost_fun), *(c->boost_e_fun));
-    };
-    white();
-}
-
-/**
- * Takes a Turncounter pointer and an integer.
- * If the count value at the pointer is 0 (counter is inactive), the turns valueis assigned.
- * @see Turncounter
- * @param c The Turncounter whose count value will be set.
- * @param turns The value to be assigned.
- */
-void setCounter(Turncounter *c, int turns)
-{
-
-    if (c->count == 0) {	// Counter is NOT already active
-        c->count = turns;
-    } else {
-        //Handle counters already activ
     }
 }
 
@@ -1146,2023 +276,6 @@ void getParams(int argc, char **argv, Fighter *player, Path *path, int optTot,
 }
 
 /**
- * Takes a Fighter pointer and an integer used to force execution.
- * If the force parameter is true, all checks are ignored.
- * If enemy's hp value is at least 50% of total, and none of atk, def or vel is 0 or less, nothing happens with an early return.
- * Otherwise, getBoost() is called to calc the level stat boost for each stat.
- * The BaseStats pointer for the fighter's fighterClass is loaded and each one of atk, def and vel is checked accounting for level boost.
- * If none of them is below the respective treshold of 35, 18 and 30 % of total, nothing happens.
- * Otherwise, all of them are reset to full amount accounting for permboosts and level boost.
- * @see Fighter
- * @see fighterClass
- * @see getBoost()
- * @param player The Fighter pointer to check the stats for.
- * @param force The integer to bypass all checks if true.
- */
-void statReset(Fighter *player, int force)
-{
-    log_tag("debug_log.txt", "[DEBUG]",
-            "Call to statReset() with ($force) == (%i)", force);
-    if (!force && (player->hp >= 0.5 * player->totalhp)
-        && !(player->atk <= 0 || player->def <= 0 || player->vel <= 0)) {
-        return;
-    }
-
-    int boost = getBoost(player->level, player->luck);
-
-    BaseStats *base = &basestats[player->class];
-    if (force || player->vel <= 0.3 * (base->vel + boost)
-        || player->atk <= 0.35 * (base->atk + boost)
-        || player->def <= 0.18 * (base->def + boost)) {
-        player->vel = base->vel + boost + player->permboost_vel;
-        player->atk = base->atk + boost + player->permboost_atk;
-        player->def = base->def + boost + player->permboost_def;
-        //Reset stats
-        if (!force) {
-            //yellow();
-            //printf("\n\n\t%s's stats reset.\n",player->name);
-            //white();
-        }
-    }
-}
-
-/**
- * Takes a Fighter and a Enemy pointers and compares their stats to determine who gets damaged and returns the fightStatus value.
- * Prints notifications to the passed WINDOW pointer.
- * On enemy death, there's a chance to call dropConsumable, dropEquip or dropArtifact (guaranteed for beast enemies).
- * @see Fighter
- * @see Enemy
- * @see statReset()
- * @see statResetEnemy()
- * @see stringFromEClass()
- * @see dropConsumable()
- * @see dropEquip()
- * @see dropArtifact()
- * @param player The Fighter pointer at hand.
- * @param e The Enemy pointer at hand.
- * @param notify_win The WINDOW pointer to call display_notification() on.
- * @param kls The Koliseo used for allocations.
- * @see display_notification()
- */
-int fight(Fighter *player, Enemy *e, WINDOW *notify_win, Koliseo *kls)
-{
-
-    fightResult res = FIGHTRES_NO_DMG;
-    char msg[200];
-    //Stat comparisons
-    //
-
-    int atkdelta = (player->atk + player->equipboost_atk) - e->atk - (rand() % 3) - 1;	//Skewed with defender
-    int defdelta = (player->def + player->equipboost_def) - e->def + (rand() % 2) + 1;	//Skewed with attacker
-    int veldelta =
-        (player->vel + player->equipboost_vel) - e->vel + (rand() % 3) + 1;
-
-    int atkOnPlayer =
-        e->atk - (player->def + player->equipboost_def + (player->vel / 6));
-    int atkOnEnemy =
-        (player->atk + player->equipboost_atk) - (e->def + (e->vel / 6));
-
-    if (G_GODMODE_ON == 1) {
-        log_tag("debug_log.txt", "[DEBUG]", "fight():  G_GODMODE_ON == 1");
-        atkdelta = 100;
-        defdelta = 100;
-        veldelta = 100;
-        atkOnPlayer = 1;
-        atkOnEnemy = 100;
-    }
-
-    int damageDealt = -1;
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "atkdelta %i", atkdelta);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "defdelta %i", defdelta);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "veldelta %i", veldelta);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "atkOnEnemy %i", atkOnEnemy);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "atkOnPlayer %i", atkOnPlayer);
-
-    if (veldelta >= 0) {
-
-        if (atkOnEnemy > 3) {
-            damageDealt = atkOnEnemy;
-            e->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result A WIN (faster, great atk).");
-        } else if (atkOnEnemy >= 0) {
-            damageDealt = abs(atkOnEnemy - atkdelta);
-            e->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result B WIN (faster, ok atk).");
-        } else {
-            if (atkOnEnemy > -3) {
-                damageDealt =
-                    fabsf(atkOnPlayer -
-                          0.75F * (player->vel + player->equipboost_vel));
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result C1 LOST (faster, atk > -3).");
-            } else {
-                damageDealt = abs(atkOnPlayer - 1);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result C2 LOST (faster, atk < -3).");
-            }
-            player->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-        }
-    } else {
-        atkdelta = -atkdelta;
-        if (atkOnPlayer > 3) {
-            damageDealt = atkOnPlayer;
-            player->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result D LOST (slower, great enemy atk).");
-        } else if (atkOnPlayer >= 0) {
-            damageDealt = abs(atkOnPlayer - atkdelta);
-            player->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result E LOST (slower, ok enemy atk).");
-        } else {
-            if (atkOnPlayer > -3) {
-                damageDealt = fabsf(atkOnEnemy - 0.75F * e->vel);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result F1 WIN (slower, enemy atk > -3).");
-            } else {
-                damageDealt = abs(atkOnEnemy - 1);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result F2 WIN (slower, enemy atk < -3).");
-            }
-            e->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-        }
-    }
-    log_tag("debug_log.txt", "[FIGHT]", "damageCalc %i", damageDealt);
-
-    int yourhit = (res == FIGHTRES_DMG_DEALT) ? 1 : 0;
-    char victim[25];
-
-    if (!yourhit) {
-
-        e->vel--;
-        e->atk--;
-        e->def -= 2;
-
-        //Check if someone earned a stat reset after the fight
-        statReset(player, 0);
-        statResetEnemy(e, 0);
-
-        strcpy(victim, player->name);
-    } else {
-
-        player->vel--;
-        player->atk--;
-        player->def -= 2;
-
-        //Account for vampirism perk
-        int vampire_perks = player->perks[VAMPIRISM]->innerValue;
-        if (vampire_perks > 0) {
-            int recovery = floor(damageDealt * (0.1 * vampire_perks));
-            player->hp += recovery;
-            log_tag("debug_log.txt", "[PERKS]", "Vampirism proc for +%i HP.",
-                    recovery);
-            if (player->hp >= player->totalhp) {
-                player->hp = player->totalhp;
-            };
-        }
-        //Account for burn on touch perk
-        int hotbody_perks = player->perks[HOT_BODY]->innerValue;
-        if (hotbody_perks > 0) {
-            int burnchance = 11 - hotbody_perks;
-            if (rand() % burnchance == 0) {
-                //TODO
-                //Handle multiple statuses
-                e->status = Burned;	//Set status to Burned. May need change to manage multiple statuses active at once
-                setCounter((Turncounter *) e->counters[Burned], 2);	//Give 2 turns of Burned status
-                log_tag("debug_log.txt", "[PERKS]",
-                        "Hotbody proc on 1/%i chance.", burnchance);
-            }
-        }
-        //Account for poison on touch perk. Order of checks with hot_body perk may cause issues?
-        int biohazard_perks = player->perks[BIOHAZARD]->innerValue;
-        if (biohazard_perks > 0) {
-            int poisonchance = 11 - biohazard_perks;
-            if (rand() % poisonchance == 0) {
-                e->status = Poison;	//Set status to Poison. May need change to manage multiple statuses active at once
-                setCounter((Turncounter *) e->counters[POISON], 2);	//Give 2 turns of Poison status
-                log_tag("debug_log.txt", "[PERKS]",
-                        "Biohazard proc on 1/%i chance.", poisonchance);
-            }
-        }
-        //Check if someone earned a stat reset after the fight
-        statResetEnemy(e, 0);
-        statReset(player, 0);
-
-        strcpy(victim, stringFromEClass(e->class));
-    }
-
-    int color = -1;
-    if (yourhit) {
-        color = S4C_WHITE;
-    } else {
-        color = S4C_RED;
-    }
-
-    sprintf(msg, "%s was hit.    (%i DMG)", victim,
-            damageDealt > 0 ? damageDealt : 1);
-    wattron(notify_win, COLOR_PAIR(color));
-    display_notification(notify_win, msg, 500);
-    wattroff(notify_win, COLOR_PAIR(color));
-
-    //Rolls
-    //
-    //Critical hit roll
-
-    //Account for critboost_chance perks
-    int critboost_value = 1.5 * player->perks[CRITBOOST_CHANCE]->innerValue;
-    int critMax = round(10.0 - floor(player->luck / 5) - (critboost_value));
-
-    int critRes = (rand() % critMax);
-
-    if (res == FIGHTRES_DMG_DEALT && (critRes <= 0)) {
-
-        //Account for critboost_dmg perks
-        int dmgboost_perks = player->perks[CRITBOOST_DMG]->innerValue;
-        damageDealt *= (0.30 + (0.12 * dmgboost_perks));
-        e->hp -= (damageDealt > 0 ? damageDealt : 1);
-        log_tag("debug_log.txt", "[FIGHT]",
-                "Critical hit for %i dmg, proc on 1/%i chance.", damageDealt,
-                critMax);
-        log_tag("debug_log.txt", "[PERKS]", "Critical hit, critboost was %i.",
-                critboost_value);
-
-        sprintf(msg, "A critical hit!    (%i DMG)",
-                damageDealt > 0 ? damageDealt : 1);
-        wattron(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        //Update stats
-        player->stats->criticalhits++;
-    }
-    //Check for deaths -> exit condition from loop
-    //
-    //
-    //
-    if (e->hp <= 0) {
-        res = FIGHTRES_KILL_DONE;
-
-        //Account for runic circle perk
-        int runic_perks = player->perks[RUNIC_MAGNET]->innerValue;
-        if (runic_perks > 0) {
-            int recovery = round(0.51 * runic_perks);
-            player->energy += recovery;
-            log_tag("debug_log.txt", "[PERKS]",
-                    "Runicmagnet proc for %i energy.", recovery);
-        }
-        if (e->beast) {
-            color = S4C_MAGENTA;
-        } else {
-            color = S4C_RED;
-        }
-        wattron(notify_win, COLOR_PAIR(color));
-        sprintf(msg, "%s fainted.", stringFromEClass(e->class));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(color));
-
-        log_tag("debug_log.txt", "[FIGHT]", "Killed  %s.",
-                stringFromEClass(e->class));
-
-        //Update stats
-        player->stats->enemieskilled++;
-    } else {
-        //Apply status effects to enemy
-        if (e->status != Normal) {
-            applyEStatus(notify_win, e);
-            log_tag("debug_log.txt", "[STATUS]", "Applied  %s to %s.",
-                    stringFromStatus(e->status), stringFromEClass(e->class));
-        }
-    }
-
-    if (player->hp <= 0) {
-        res = FIGHTRES_DEATH;
-    } else {
-        //Apply status effects to player
-        if (player->status != Normal) {
-            applyStatus(notify_win, player);
-        }
-    }
-
-    //Consumable drop, guaranteed on killing a beast
-    if (res == FIGHTRES_KILL_DONE
-        && (e->beast || ((rand() % 9) - (player->luck / 10) <= 0))) {
-        int drop = dropConsumable(player);
-        sprintf(msg, "You found a %s!", stringFromConsumables(drop));
-        wattron(notify_win, COLOR_PAIR(S4C_CYAN));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_CYAN));
-        log_tag("debug_log.txt", "[DROPS]", "Found Consumable:    %s.",
-                stringFromConsumables(drop));
-    }
-
-    //Artifact drop (if we don't have all of them), guaranteed on killing a beast
-    if ((player->stats->artifactsfound != ARTIFACTSMAX + 1)
-        && res == FIGHTRES_KILL_DONE && (e->beast
-                                         ||
-                                         ((rand() % ENEMY_ARTIFACTDROP_CHANCE) -
-                                          (player->luck / 10) <= 0))) {
-        int artifact_drop = dropArtifact(player);
-        sprintf(msg, "You found a %s!", stringFromArtifacts(artifact_drop));
-        wattron(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        log_tag("debug_log.txt", "[DROPS]", "Found Artifact:    %s.",
-                stringFromArtifacts(artifact_drop));
-        if (!e->beast)
-            log_tag("debug_log.txt", "[.1%% CHANCE]",
-                    "\nNORMAL ENEMY DROPPED ARTIFACT! 0.1%% chance??\n");
-    }
-
-    //Equip drop, guaranteed on killing a beast
-    if (res == FIGHTRES_KILL_DONE
-        && (e->beast || ((rand() % 15) - (player->luck / 10) <= 0))) {
-        dropEquip(player, e->beast, notify_win, kls);
-    }
-    return res;
-}
-
-/**
- * Takes an Enemy and a Fighter pointer and compares their stats to determine who gets damaged and returns the fightStatus value.
- * Prints notifications to the passed WINDOW pointer.
- * On enemy death, there's a chance to call dropConsumable, dropEquip or dropArtifact (guaranteed for beast enemies).
- * NOTE:  that the return values are always from the POV of the Fighter: FIGHTRES_DMG_DEALT means the Enemy was damaged!
- * @see defer_fight_enemy()
- * @see Fighter
- * @see Enemy
- * @see statReset()
- * @see statResetEnemy()
- * @see stringFromEClass()
- * @see dropConsumable()
- * @see dropEquip()
- * @see dropArtifact()
- * @param player The Fighter pointer at hand.
- * @param e The Enemy pointer at hand.
- * @param notify_win The WINDOW pointer to call display_notification() on.
- * @param kls The Koliseo used for allocations.
- * @see display_notification()
- */
-int enemy_attack(Enemy *e, Fighter *target, WINDOW *notify_win, Koliseo *kls)
-{
-    //Implementation similar to fight(), as a base idea
-    //Should return fightResult values, while keeping the perspective on the Fighter, as in:
-    //
-    //      FIGHTRES_DEATH means the Fighter died
-    //      FIGHTRES_KILL_DONE means the Enemy died
-    //      FIGHTRES_DMG_DEALT means the Fighter inflicted damage
-    //      FIGHRES_DMG_TAKEN means the Fighter received damage
-
-    fightResult res = FIGHTRES_NO_DMG;
-    char msg[200];
-    //Stat comparisons
-    //
-
-    int atkdelta = (e->atk + e->turnboost_atk) - (target->atk + target->equipboost_atk + target->turnboost_atk - (rand() % 3)) - 1;	//Skewed with defender
-    int defdelta = (e->def + e->turnboost_def) - (target->def + target->equipboost_def + target->turnboost_def) + (rand() % 2) + 1;	//Skewed with attacker
-    int veldelta =
-        (e->vel + e->turnboost_vel) - (target->vel + target->equipboost_vel +
-                                       target->turnboost_vel) + (rand() % 3) +
-        1;
-
-    int atkOnPlayer =
-        (e->atk + e->turnboost_atk) - (target->def + target->equipboost_def +
-                                       target->turnboost_def +
-                                       (target->vel / 6));
-    int atkOnEnemy =
-        (target->atk + target->equipboost_atk + target->turnboost_atk) -
-        (e->def + e->turnboost_def + (e->vel / 6));
-
-    if (G_GODMODE_ON == 1) {
-        log_tag("debug_log.txt", "[DEBUG]", "[%s]:  G_GODMODE_ON == 1",
-                __func__);
-        atkdelta = -100;
-        defdelta = -100;
-        veldelta = -100;
-        atkOnPlayer = 1;
-        atkOnEnemy = 100;
-    }
-
-    int damageDealt = -1;
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "atkdelta %i", atkdelta);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "defdelta %i", defdelta);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "veldelta %i", veldelta);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "atkOnEnemy %i", atkOnEnemy);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "atkOnPlayer %i", atkOnPlayer);
-
-    if (veldelta >= 0) {	//Enemy has a non-negative veldelta
-        if (atkOnPlayer > 3) {
-            damageDealt = atkOnPlayer;
-            target->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "[%s]:  Fight result D LOST (slower, great enemy atk).",
-                    __func__);
-        } else if (atkOnPlayer >= 0) {
-            damageDealt = abs(atkOnPlayer - atkdelta);
-            target->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "[%s]:  Fight result E LOST (slower, ok enemy atk).",
-                    __func__);
-        } else {
-            if (atkOnPlayer > -3) {
-                damageDealt = fabsf(atkOnEnemy - 0.75F * e->vel);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "[%s]:  Fight result F1 WIN (slower, enemy atk > -3).",
-                        __func__);
-            } else {
-                damageDealt = abs(atkOnEnemy - 1);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "[%s]:  Fight result F2 WIN (slower, enemy atk < -3).",
-                        __func__);
-            }
-            e->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-        }
-    } else {			//Enemy veldelta is not strictly positive
-        atkdelta = -atkdelta;
-        if (atkOnEnemy > 3) {
-            damageDealt = atkOnEnemy;
-            e->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "[%s]:  Fight result A WIN (faster, great atk).", __func__);
-        } else if (atkOnEnemy >= 0) {
-            damageDealt = abs(atkOnEnemy - atkdelta);
-            e->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "[%s]:  Fight result B WIN (faster, ok atk).", __func__);
-        } else {
-            if (atkOnEnemy > -3) {
-                damageDealt =
-                    fabsf(atkOnPlayer -
-                          0.75F * (target->vel + target->equipboost_vel));
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "[%s]:  Fight result C1 LOST (faster, atk > -3).",
-                        __func__);
-            } else {
-                damageDealt = abs(atkOnPlayer - 1);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "[%s]:  Fight result C2 LOST (faster, atk < -3).",
-                        __func__);
-            }
-            target->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-        }
-    }
-    log_tag("debug_log.txt", "[FIGHT]", "[%s]:  damageCalc %i", __func__,
-            damageDealt);
-
-    int playerhit = (res == FIGHTRES_DMG_DEALT) ? 1 : 0;
-    char victim[25];
-
-    if (!playerhit) {
-
-        e->vel--;
-        e->atk--;
-        e->def -= 2;
-
-        //Check if someone earned a stat reset after the fight
-        statReset(target, 0);
-        statResetEnemy(e, 0);
-
-        strcpy(victim, target->name);
-    } else {
-
-        target->vel--;
-        target->atk--;
-        target->def -= 2;
-
-        //Account for vampirism perk
-        int vampire_perks = target->perks[VAMPIRISM]->innerValue;
-        if (vampire_perks > 0) {
-            int recovery = floor(damageDealt * (0.1 * vampire_perks));
-            target->hp += recovery;
-            log_tag("debug_log.txt", "[PERKS]", "Vampirism proc for +%i HP.",
-                    recovery);
-            if (target->hp >= target->totalhp) {
-                target->hp = target->totalhp;
-            };
-        }
-        //Account for burn on touch perk
-        int hotbody_perks = target->perks[HOT_BODY]->innerValue;
-        if (hotbody_perks > 0) {
-            int burnchance = 11 - hotbody_perks;
-            if (rand() % burnchance == 0) {
-                //TODO
-                //Handle multiple statuses
-                e->status = Burned;	//Set status to Burned. May need change to manage multiple statuses active at once
-                setCounter((Turncounter *) e->counters[Burned], 2);	//Give 2 turns of Burned status
-                log_tag("debug_log.txt", "[PERKS]",
-                        "Hotbody proc on 1/%i chance.", burnchance);
-            }
-        }
-        //Account for poison on touch perk. Order of checks with hot_body perk may cause issues?
-        int biohazard_perks = target->perks[BIOHAZARD]->innerValue;
-        if (biohazard_perks > 0) {
-            int poisonchance = 11 - biohazard_perks;
-            if (rand() % poisonchance == 0) {
-                e->status = Poison;	//Set status to Poison. May need change to manage multiple statuses active at once
-                setCounter((Turncounter *) e->counters[POISON], 2);	//Give 2 turns of Poison status
-                log_tag("debug_log.txt", "[PERKS]",
-                        "Biohazard proc on 1/%i chance.", poisonchance);
-            }
-        }
-        //Check if someone earned a stat reset after the fight
-        statResetEnemy(e, 0);
-        statReset(target, 0);
-
-        strcpy(victim, stringFromEClass(e->class));
-    }
-
-    int color = -1;
-    if (playerhit) {
-        color = S4C_WHITE;
-    } else {
-        color = S4C_RED;
-    }
-
-    sprintf(msg, "%s was hit.    (%i DMG)", victim,
-            damageDealt > 0 ? damageDealt : 1);
-    wattron(notify_win, COLOR_PAIR(color));
-    display_notification(notify_win, msg, 500);
-    wattroff(notify_win, COLOR_PAIR(color));
-
-    //Rolls
-    //
-    //Critical hit roll
-
-    //Account for critboost_chance perks
-    int critboost_value = 1.5 * target->perks[CRITBOOST_CHANCE]->innerValue;
-    int critMax = round(10.0 - floor(target->luck / 5) - (critboost_value));
-
-    int critRes = (rand() % critMax);
-
-    if (res == FIGHTRES_DMG_DEALT && (critRes <= 0)) {
-
-        //Account for critboost_dmg perks
-        int dmgboost_perks = target->perks[CRITBOOST_DMG]->innerValue;
-        damageDealt *= (0.30 + (0.12 * dmgboost_perks));
-        e->hp -= (damageDealt > 0 ? damageDealt : 1);
-        log_tag("debug_log.txt", "[FIGHT]",
-                "Critical hit for %i dmg, proc on 1/%i chance.", damageDealt,
-                critMax);
-        log_tag("debug_log.txt", "[PERKS]", "Critical hit, critboost was %i.",
-                critboost_value);
-
-        sprintf(msg, "A critical hit!    (%i DMG)",
-                damageDealt > 0 ? damageDealt : 1);
-        wattron(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        //Update stats
-        target->stats->criticalhits++;
-    }
-    //Check for deaths -> exit condition from loop
-    //
-    //
-    //
-    if (e->hp <= 0) {
-        res = FIGHTRES_KILL_DONE;
-
-        //Account for runic circle perk
-        int runic_perks = target->perks[RUNIC_MAGNET]->innerValue;
-        if (runic_perks > 0) {
-            int recovery = round(0.51 * runic_perks);
-            target->energy += recovery;
-            log_tag("debug_log.txt", "[PERKS]",
-                    "Runicmagnet proc for %i energy.", recovery);
-        }
-        if (e->beast) {
-            color = S4C_MAGENTA;
-        } else {
-            color = S4C_RED;
-        }
-        wattron(notify_win, COLOR_PAIR(color));
-        sprintf(msg, "%s fainted.", stringFromEClass(e->class));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(color));
-
-        log_tag("debug_log.txt", "[FIGHT]", "Killed  %s.",
-                stringFromEClass(e->class));
-
-        //Update stats
-        target->stats->enemieskilled++;
-    } else {
-        //Apply status effects to enemy
-        if (e->status != Normal) {
-            applyEStatus(notify_win, e);
-            log_tag("debug_log.txt", "[STATUS]", "Applied  %s to %s.",
-                    stringFromStatus(e->status), stringFromEClass(e->class));
-        }
-    }
-
-    if (target->hp <= 0) {
-        log_tag("debug_log.txt", "[DEBUG]", "[%s]:  Target died. Hp: (%i)",
-                __func__, target->hp);
-        res = FIGHTRES_DEATH;
-    } else {
-        //Apply status effects to target
-        if (target->status != Normal) {
-            applyStatus(notify_win, target);
-        }
-    }
-
-    //Consumable drop, guaranteed on killing a beast
-    if (res == FIGHTRES_KILL_DONE
-        && (e->beast || ((rand() % 9) - (target->luck / 10) <= 0))) {
-        int drop = dropConsumable(target);
-        sprintf(msg, "You found a %s!", stringFromConsumables(drop));
-        wattron(notify_win, COLOR_PAIR(S4C_CYAN));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_CYAN));
-        log_tag("debug_log.txt", "[DROPS]", "Found Consumable:    %s.",
-                stringFromConsumables(drop));
-    }
-
-    //Artifact drop (if we don't have all of them), guaranteed on killing a beast
-    if ((target->stats->artifactsfound != ARTIFACTSMAX + 1)
-        && res == FIGHTRES_KILL_DONE && (e->beast
-                                         ||
-                                         ((rand() % ENEMY_ARTIFACTDROP_CHANCE) -
-                                          (target->luck / 10) <= 0))) {
-        int artifact_drop = dropArtifact(target);
-        sprintf(msg, "You found a %s!", stringFromArtifacts(artifact_drop));
-        wattron(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        log_tag("debug_log.txt", "[DROPS]", "Found Artifact:    %s.",
-                stringFromArtifacts(artifact_drop));
-        if (!e->beast)
-            log_tag("debug_log.txt", "[.1%% CHANCE]",
-                    "\nNORMAL ENEMY DROPPED ARTIFACT! 0.1%% chance??\n");
-    }
-
-    //Equip drop, guaranteed on killing a beast
-    if (res == FIGHTRES_KILL_DONE
-        && (e->beast || ((rand() % 15) - (target->luck / 10) <= 0))) {
-        dropEquip(target, e->beast, notify_win, kls);
-    }
-    return res;
-}
-
-/**
- * Takes a Fighter and a Enemy pointers and calls fight().
- * @see Fighter
- * @see Enemy
- * @see fight()
- * @param player The Fighter pointer at hand.
- * @param e The Enemy pointer at hand.
- * @param foe_op The foeTurnOption_OP for the foe.
- * @param notify_win The WINDOW pointer to call display_notification() on.
- * @param kls The Koliseo used for allocations.
- */
-int defer_fight_enemy(Fighter *player, Enemy *e, foeTurnOption_OP foe_op,
-                      WINDOW *notify_win, Koliseo *kls)
-{
-    char msg[200];
-    //FIXME
-    //Is it okay to return just one result, when having 2 interactions that could go differently?
-    //
-    //Use FIGHTRES_CLASH as needed, to indicate both sides were damaged at some point.
-    fightResult res = FIGHTRES_NO_DMG;
-
-    int player_goes_first = (player->vel >= e->vel ? 1 : 0);
-
-    int first_act_res = FIGHTRES_NO_DMG;
-
-    if (player_goes_first) {
-
-        res = fight(player, e, notify_win, kls);
-
-        //Check res and apply second action if needed
-        log_tag("debug_log.txt", "[DEBUG]",
-                "[%s()]: First act res was [%s]: [%i]", __func__,
-                stringFrom_fightResult(res), res);
-        first_act_res = res;
-
-        if (res != FIGHTRES_DEATH && res != FIGHTRES_KILL_DONE) {
-            switch (foe_op) {
-            case FOE_OP_INVALID: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "foe_op was FOE_OP_INVALID in [%s]: [%i]", __func__,
-                        foe_op);
-                kls_free(default_kls);
-                kls_free(temporary_kls);
-                exit(EXIT_FAILURE);
-            }
-            break;
-            case FOE_OP_IDLE: {
-                log_tag("debug_log.txt", "[DEFER]",
-                        "[%s()]:  Foe { %s } was idle.", __func__,
-                        stringFromEClass(e->class));
-                wattron(notify_win, COLOR_PAIR(S4C_GREY));
-                sprintf(msg, "%s is loafing around.",
-                        stringFromEClass(e->class));
-                display_notification(notify_win, msg, 500);
-                wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-            }
-            break;
-            case FOE_OP_FIGHT: {
-                log_tag("debug_log.txt", "[DEFER]",
-                        "[%s()]:  Foe { %s } wants to fight.", __func__,
-                        stringFromEClass(e->class));
-                wattron(notify_win, COLOR_PAIR(S4C_GREY));
-                sprintf(msg, "%s is angry!", stringFromEClass(e->class));
-                display_notification(notify_win, msg, 500);
-                wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-
-                res = enemy_attack(e, player, notify_win, kls);
-            }
-            break;
-            case FOE_OP_SPECIAL: {
-                log_tag("debug_log.txt", "[TODO]",
-                        "[%s()]:  Foe { %s } wants to use a special.",
-                        __func__, stringFromEClass(e->class));
-                //TODO
-                //Implement enemy special function
-                //res = enemy_attack_special(e,player,notify_win,kls);
-            }
-            break;
-            default: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "Unexpected foeTurnOption_OP in [%s()]: [%i]",
-                        __func__, foe_op);
-                kls_free(default_kls);
-                kls_free(temporary_kls);
-                exit(EXIT_FAILURE);
-            }
-            break;
-            }			// End foe_op switch
-
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "[%s()]: Second act res was [%s]: [%i]", __func__,
-                    stringFrom_fightResult(res), res);
-            if (res == FIGHTRES_DEATH || res == FIGHTRES_KILL_DONE) {
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "[%s()]: Deferred fight was not a clash...", __func__);
-            } else if ((res == FIGHTRES_DMG_TAKEN
-                        && first_act_res == FIGHTRES_DMG_DEALT)
-                       || (res == FIGHTRES_DMG_DEALT
-                           && first_act_res == FIGHTRES_DMG_TAKEN)) {
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "[%s()]: Deferred fight was a clash!", __func__);
-                res = FIGHTRES_CLASH;
-            }
-
-            return res;
-        } else if (res == FIGHTRES_DEATH) {
-            return res;
-        } else if (res == FIGHTRES_KILL_DONE) {
-            return res;
-        }
-    } else {
-        //Foe acts first
-        switch (foe_op) {
-        case FOE_OP_INVALID: {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "foe_op was FOE_OP_INVALID in [%s]: [%i]", __func__,
-                    foe_op);
-            kls_free(default_kls);
-            kls_free(temporary_kls);
-            exit(EXIT_FAILURE);
-        }
-        break;
-        case FOE_OP_IDLE: {
-            log_tag("debug_log.txt", "[DEFER]",
-                    "[%s()]:  Foe { %s } was idle.", __func__,
-                    stringFromEClass(e->class));
-            wattron(notify_win, COLOR_PAIR(S4C_GREY));
-            sprintf(msg, "%s is loafing around.",
-                    stringFromEClass(e->class));
-            display_notification(notify_win, msg, 500);
-            wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-        }
-        break;
-        case FOE_OP_FIGHT: {
-            log_tag("debug_log.txt", "[DEFER]",
-                    "[%s()]:  Foe { %s } wants to fight.", __func__,
-                    stringFromEClass(e->class));
-            res = enemy_attack(e, player, notify_win, kls);
-        }
-        break;
-        case FOE_OP_SPECIAL: {
-            log_tag("debug_log.txt", "[TODO]",
-                    "[%s()]:  Foe { %s } wants to use a special.", __func__,
-                    stringFromEClass(e->class));
-            //TODO
-            //Implement enemy special function
-            //res = enemy_attack_special(e,player,notify_win,kls);
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Unexpected foeTurnOption_OP in [%s()]: [%i]", __func__,
-                    foe_op);
-            kls_free(default_kls);
-            kls_free(temporary_kls);
-            exit(EXIT_FAILURE);
-        }
-        break;
-        }			// End foe_op switch
-
-        //Check res and apply second action if needed
-        log_tag("debug_log.txt", "[DEBUG]",
-                "[%s()]: First act res was [%s]: [%i]", __func__,
-                stringFrom_fightResult(res), res);
-        first_act_res = res;
-
-        if (res != FIGHTRES_DEATH && res != FIGHTRES_KILL_DONE) {
-            res = fight(player, e, notify_win, kls);
-
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "[%s()]: Second act res was [%s]: [%i]", __func__,
-                    stringFrom_fightResult(res), res);
-            if (res == FIGHTRES_DEATH || res == FIGHTRES_KILL_DONE) {
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "[%s()]: Deferred fight was not a clash...", __func__);
-            } else if ((res == FIGHTRES_DMG_TAKEN
-                        && first_act_res == FIGHTRES_DMG_DEALT)
-                       || (res == FIGHTRES_DMG_DEALT
-                           && first_act_res == FIGHTRES_DMG_TAKEN)) {
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "[%s()]: Deferred fight was a clash!", __func__);
-                res = FIGHTRES_CLASH;
-            }
-
-            return res;
-        } else if (res == FIGHTRES_DEATH) {
-            return res;
-        } else if (res == FIGHTRES_KILL_DONE) {
-            return res;
-        }
-    }
-    return res;
-}
-
-/**
- * Takes a Fighter and a Enemy pointers and calls do_Skill().
- * @see Fighter
- * @see Enemy
- * @see do_Skill()
- * @param player The Fighter pointer at hand.
- * @param e The Enemy pointer at hand.
- * @param picked_skill The picked skill by Fighter.
- * @param foe_op The foeTurnOption_OP for the foe.
- * @param notify_win The WINDOW pointer to call display_notification() on.
- * @param kls The Koliseo used for allocations.
- */
-int defer_skill_enemy(Fighter *player, Enemy *e, skillType picked_skill, foeTurnOption_OP foe_op,
-                      WINDOW *notify_win, Koliseo *kls)
-{
-    char msg[200];
-    //FIXME
-    //Is it okay to return just one result, when having 2 interactions that could go differently?
-    //
-    //Use FIGHTRES_CLASH as needed, to indicate both sides were damaged at some point.
-    fightResult res = FIGHTRES_NO_DMG;
-
-    int player_goes_first = (player->vel >= e->vel ? 1 : 0);
-
-    int first_act_res = FIGHTRES_NO_DMG;
-
-    if (player_goes_first) {
-
-        res = do_Skill(player, e, picked_skill, notify_win, kls);
-
-        //Check res and apply second action if needed
-        log_tag("debug_log.txt", "[DEBUG]",
-                "[%s()]: First act res was [%s]: [%i]", __func__,
-                stringFrom_fightResult(res), res);
-        first_act_res = res;
-
-        if (res != FIGHTRES_DEATH && res != FIGHTRES_KILL_DONE) {
-            switch (foe_op) {
-            case FOE_OP_INVALID: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "foe_op was FOE_OP_INVALID in [%s]: [%i]", __func__,
-                        foe_op);
-                kls_free(default_kls);
-                kls_free(temporary_kls);
-                exit(EXIT_FAILURE);
-            }
-            break;
-            case FOE_OP_IDLE: {
-                log_tag("debug_log.txt", "[DEFER]",
-                        "[%s()]:  Foe { %s } was idle.", __func__,
-                        stringFromEClass(e->class));
-                wattron(notify_win, COLOR_PAIR(S4C_GREY));
-                sprintf(msg, "%s is loafing around.",
-                        stringFromEClass(e->class));
-                display_notification(notify_win, msg, 500);
-                wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-            }
-            break;
-            case FOE_OP_FIGHT: {
-                log_tag("debug_log.txt", "[DEFER]",
-                        "[%s()]:  Foe { %s } wants to fight.", __func__,
-                        stringFromEClass(e->class));
-                wattron(notify_win, COLOR_PAIR(S4C_GREY));
-                sprintf(msg, "%s is angry!", stringFromEClass(e->class));
-                display_notification(notify_win, msg, 500);
-                wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-
-                res = enemy_attack(e, player, notify_win, kls);
-            }
-            break;
-            case FOE_OP_SPECIAL: {
-                log_tag("debug_log.txt", "[TODO]",
-                        "[%s()]:  Foe { %s } wants to use a special.",
-                        __func__, stringFromEClass(e->class));
-                //TODO
-                //Implement enemy special function
-                //res = enemy_attack_special(e,player,notify_win,kls);
-            }
-            break;
-            default: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "Unexpected foeTurnOption_OP in [%s()]: [%i]",
-                        __func__, foe_op);
-                kls_free(default_kls);
-                kls_free(temporary_kls);
-                exit(EXIT_FAILURE);
-            }
-            break;
-            }			// End foe_op switch
-
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "[%s()]: Second act res was [%s]: [%i]", __func__,
-                    stringFrom_fightResult(res), res);
-            if (res == FIGHTRES_DEATH || res == FIGHTRES_KILL_DONE) {
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "[%s()]: Deferred fight was not a clash...", __func__);
-            } else if ((res == FIGHTRES_DMG_TAKEN
-                        && first_act_res == FIGHTRES_DMG_DEALT)
-                       || (res == FIGHTRES_DMG_DEALT
-                           && first_act_res == FIGHTRES_DMG_TAKEN)) {
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "[%s()]: Deferred fight was a clash!", __func__);
-                res = FIGHTRES_CLASH;
-            }
-
-            return res;
-        } else if (res == FIGHTRES_DEATH) {
-            return res;
-        } else if (res == FIGHTRES_KILL_DONE) {
-            return res;
-        }
-    } else {
-        //Foe acts first
-        switch (foe_op) {
-        case FOE_OP_INVALID: {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "foe_op was FOE_OP_INVALID in [%s]: [%i]", __func__,
-                    foe_op);
-            kls_free(default_kls);
-            kls_free(temporary_kls);
-            exit(EXIT_FAILURE);
-        }
-        break;
-        case FOE_OP_IDLE: {
-            log_tag("debug_log.txt", "[DEFER]",
-                    "[%s()]:  Foe { %s } was idle.", __func__,
-                    stringFromEClass(e->class));
-            wattron(notify_win, COLOR_PAIR(S4C_GREY));
-            sprintf(msg, "%s is loafing around.",
-                    stringFromEClass(e->class));
-            display_notification(notify_win, msg, 500);
-            wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-        }
-        break;
-        case FOE_OP_FIGHT: {
-            log_tag("debug_log.txt", "[DEFER]",
-                    "[%s()]:  Foe { %s } wants to fight.", __func__,
-                    stringFromEClass(e->class));
-            res = enemy_attack(e, player, notify_win, kls);
-        }
-        break;
-        case FOE_OP_SPECIAL: {
-            log_tag("debug_log.txt", "[TODO]",
-                    "[%s()]:  Foe { %s } wants to use a special.", __func__,
-                    stringFromEClass(e->class));
-            //TODO
-            //Implement enemy special function
-            //res = enemy_attack_special(e,player,notify_win,kls);
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Unexpected foeTurnOption_OP in [%s()]: [%i]", __func__,
-                    foe_op);
-            kls_free(default_kls);
-            kls_free(temporary_kls);
-            exit(EXIT_FAILURE);
-        }
-        break;
-        }			// End foe_op switch
-
-        //Check res and apply second action if needed
-        log_tag("debug_log.txt", "[DEBUG]",
-                "[%s()]: First act res was [%s]: [%i]", __func__,
-                stringFrom_fightResult(res), res);
-        first_act_res = res;
-
-        if (res != FIGHTRES_DEATH && res != FIGHTRES_KILL_DONE) {
-            res = do_Skill(player, e, picked_skill, notify_win, kls);
-
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "[%s()]: Second act res was [%s]: [%i]", __func__,
-                    stringFrom_fightResult(res), res);
-            if (res == FIGHTRES_DEATH || res == FIGHTRES_KILL_DONE) {
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "[%s()]: Deferred fight was not a clash...", __func__);
-            } else if ((res == FIGHTRES_DMG_TAKEN
-                        && first_act_res == FIGHTRES_DMG_DEALT)
-                       || (res == FIGHTRES_DMG_DEALT
-                           && first_act_res == FIGHTRES_DMG_TAKEN)) {
-                log_tag("debug_log.txt", "[DEBUG]",
-                        "[%s()]: Deferred fight was a clash!", __func__);
-                res = FIGHTRES_CLASH;
-            }
-
-            return res;
-        } else if (res == FIGHTRES_DEATH) {
-            return res;
-        } else if (res == FIGHTRES_KILL_DONE) {
-            return res;
-        }
-    }
-    return res;
-}
-
-/**
- * Takes a Fighter, a Boss and a Path pointers and compares fighters stats to determine who gets damaged and returns the fightStatus value.
- * Prints notifications to the passed WINDOW pointer.
- * On boss death, we call dropConsumable, dropEquip and dropArtifact.
- * @see Fighter
- * @see Boss
- * @see statReset()
- * @see statResetBoss()
- * @see stringFromBossClass()
- * @see dropConsumable()
- * @see dropEquip()
- * @see dropArtifact()
- * @param player The Fighter pointer at hand.
- * @param b The Enemy pointer at hand.
- * @param p The Path pointer for the game.
- * @param notify_win The WINDOW pointer to call display_notification() on.
- * @param kls The Koliseo used for allocations.
- * @see display_notification()
- */
-int boss_fight(Fighter *player, Boss *b, Path *p, WINDOW *notify_win,
-               Koliseo *kls)
-{
-
-    fightResult res = FIGHTRES_NO_DMG;
-    //Stat comparisons
-    //
-    char msg[200];
-
-    int atkdelta = (player->atk + player->equipboost_atk) - b->atk - (rand() % 3) - 1;	//Skewed with defender
-    int defdelta = (player->def + player->equipboost_def) - b->def + (rand() % 2) + 1;	//Skewed with attacker
-    int veldelta =
-        (player->vel + player->equipboost_vel) - b->vel + (rand() % 3) + 1;
-
-    int atkOnPlayer =
-        b->atk - (player->def + player->equipboost_def + (player->vel / 6));
-    int atkOnEnemy =
-        (player->atk + player->equipboost_atk) - (b->def + (b->vel / 6));
-    if (G_GODMODE_ON == 1) {
-        log_tag("debug_log.txt", "[DEBUG]", "boss_fight():  G_GODMODE_ON == 1");
-        atkdelta = 100;
-        defdelta = 100;
-        veldelta = 100;
-        atkOnPlayer = 1;
-        atkOnEnemy = 100;
-    }
-    sprintf(msg, "atkdelta %i", atkdelta);
-    log_tag("debug_log.txt", "[FIGHT-BOSS]", msg);
-    sprintf(msg, "defdelta %i", defdelta);
-    log_tag("debug_log.txt", "[FIGHT-BOSS]", msg);
-    sprintf(msg, "veldelta %i", veldelta);
-    log_tag("debug_log.txt", "[FIGHT-BOSS]", msg);
-    sprintf(msg, "atkOnEnemy %i", atkOnEnemy);
-    log_tag("debug_log.txt", "[FIGHT-BOSS]", msg);
-    sprintf(msg, "atkOnPlayer %i\n", atkOnPlayer);
-    log_tag("debug_log.txt", "[FIGHT-BOSS]", msg);
-
-    int damageDealt = -1;
-
-    if (veldelta >= 0) {
-
-        if (atkOnEnemy > 3) {
-            damageDealt = atkOnEnemy;
-            b->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result A WIN (faster, great atk).\n");
-        } else if (atkOnEnemy >= 0) {
-            damageDealt = abs(atkOnEnemy - atkdelta);
-            b->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result B WIN (faster, ok atk).\n");
-        } else {
-            if (atkOnEnemy < -3) {
-                damageDealt =
-                    fabsf(atkOnPlayer -
-                          0.75F * (player->vel + player->equipboost_vel));
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result C1 LOST (faster, atk < -3).\n");
-            } else {
-                damageDealt = abs(atkOnPlayer - 1);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result C2 LOST (faster, atk > -3).\n");
-            }
-            player->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-        }
-    } else {
-        atkdelta = -atkdelta;
-        if (atkOnPlayer > 3) {
-            damageDealt = atkOnPlayer;
-            player->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result D LOST (slower, great enemy atk.\n");
-        } else if (atkOnPlayer >= 0) {
-            damageDealt = abs(atkOnPlayer - atkdelta);
-            player->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result E LOST (slower, ok enemy atk.\n");
-        } else {
-            if (atkOnPlayer < -3) {
-                damageDealt = fabsf(atkOnEnemy - 0.75F * b->vel);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result F1 WIN (slower, enemy atk < -3.\n");
-            } else {
-                damageDealt = abs(atkOnEnemy - 1);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result F2 WIN (slower, enemy atk > -3.\n");
-            }
-            b->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-        }
-    }
-
-    log_tag("debug_log.txt", "[FIGHT]", "damageCalc %i", damageDealt);
-
-    int yourhit = (res == FIGHTRES_DMG_DEALT) ? 1 : 0;
-    char victim[25];
-
-    if (!yourhit) {
-
-        b->vel--;
-        b->atk--;
-        b->def -= 2;
-
-        //Check if someone earned a stat reset after the fight
-        statReset(player, 0);
-        statResetBoss(b, 0);
-
-        strcpy(victim, player->name);
-    } else {
-
-        player->vel--;
-        player->atk--;
-        player->def -= 2;
-
-        //Account for vampirism perk
-        int vampire_perks = player->perks[VAMPIRISM]->innerValue;
-        if (vampire_perks > 0) {
-            int recovery = floor(damageDealt * (0.1 * vampire_perks));
-            player->hp += recovery;
-            sprintf(msg, "Vampirism proc for +%i HP.\n", recovery);
-            log_tag("debug_log.txt", "[PERKS]", msg);
-            if (player->hp >= player->totalhp) {
-                player->hp = player->totalhp;
-            };
-        }
-        //Account for burn on touch perk
-        int hotbody_perks = player->perks[HOT_BODY]->innerValue;
-        if (hotbody_perks > 0) {
-            int burnchance = 11 - hotbody_perks;
-            if (rand() % burnchance == 0) {
-                b->status = Burned;	//Set status to Burned. May need change to manage multiple statuses active at once
-                setCounter((Turncounter *) b->counters[Burned], 2);	//Give 2 turns of Burned status
-                sprintf(msg, "Hotbody proc on 1/%i chance.\n", burnchance);
-                log_tag("debug_log.txt", "[PERKS]", msg);
-            }
-        }
-        //Account for poison on touch perk. Order of checks with hot_body perk may cause issues?
-        int biohazard_perks = player->perks[BIOHAZARD]->innerValue;
-        if (biohazard_perks > 0) {
-            int poisonchance = 11 - biohazard_perks;
-            if (rand() % poisonchance == 0) {
-                b->status = Poison;	//Set status to Poison. May need change to manage multiple statuses active at once
-                setCounter((Turncounter *) b->counters[POISON], 2);	//Give 2 turns of Poison status
-                sprintf(msg, "Biohazard proc on 1/%i chance.\n", poisonchance);
-                log_tag("debug_log.txt", "[PERKS]", msg);
-            }
-        }
-        //Check if someone earned a stat reset after the fight
-        statResetBoss(b, 0);
-        statReset(player, 0);
-
-        strcpy(victim, stringFromBossClass(b->class));
-    }
-
-    int color = -1;
-    if (yourhit) {
-        color = S4C_WHITE;
-    } else {
-        color = S4C_RED;
-    }
-
-    wattron(notify_win, COLOR_PAIR(color));
-    sprintf(msg, "%s was hit.    (%i DMG)", victim,
-            damageDealt > 0 ? damageDealt : 1);
-    display_notification(notify_win, msg, 500);
-    wattroff(notify_win, COLOR_PAIR(color));
-
-    //Rolls
-    //
-    //Critical hit roll
-
-    //Account for critboost_chance perks
-    int critMax =
-        round(10.0 - floor(player->luck / 5) -
-              (1.5 * player->perks[CRITBOOST_CHANCE]->innerValue));
-
-    int critRes = (rand() % critMax);
-
-    if (res == FIGHTRES_DMG_DEALT && (critRes <= 0)) {
-
-        //Account for critboost_dmg perks
-        int dmgboost_perks = player->perks[CRITBOOST_DMG]->innerValue;
-        damageDealt *= (0.30 + (0.12 * dmgboost_perks));
-        b->hp -= damageDealt > 0 ? damageDealt : 1;
-        wattron(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        sprintf(msg, "A critical hit!    (%i DMG)",
-                damageDealt > 0 ? damageDealt : 1);
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        sprintf(msg, "Critical hit for %i dmg, proc on 1/%i chance.\n",
-                damageDealt, critMax);
-        log_tag("debug_log.txt", "[FIGHT-BOSS]", msg);
-
-        //Update stats
-        player->stats->criticalhits++;
-    }
-    //Check for deaths -> exit condition from loop
-    //
-    //
-    //
-    if (b->hp <= 0) {
-        res = FIGHTRES_KILL_DONE;
-
-        //Account for runic circle perk
-        int runic_perks = player->perks[RUNIC_MAGNET]->innerValue;
-        if (runic_perks > 0) {
-            int recovery = round(0.51 * runic_perks);
-            player->energy += recovery;
-            sprintf(msg, "Runicmagnet proc for %i energy.\n", recovery);
-            log_tag("debug_log.txt", "[PERKS]", msg);
-        }
-        if (b->beast) {
-            color = S4C_MAGENTA;
-        } else {
-            color = S4C_RED;
-        }
-        wattron(notify_win, COLOR_PAIR(color));
-        sprintf(msg, "%s fainted.", stringFromBossClass(b->class));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(color));
-        sprintf(msg, "Killed  %s.", stringFromBossClass(b->class));
-        log_tag("debug_log.txt", "[FIGHT-BOSS]", msg);
-
-        //Update stats
-        player->stats->bosseskilled++;
-
-        //Check if we killed this boss before
-        if (player->stats->killed_bosses[b->class] == 0) {
-            //Update stats
-            player->stats->unique_bosseskilled++;
-            player->stats->killed_bosses[b->class] += 1;
-            //Check if we have to update the wincon value
-            if (p->win_condition->class == ALL_BOSSES) {
-                p->win_condition->current_val++;
-            }
-            sprintf(msg, "Killed new boss  %s.\n",
-                    stringFromBossClass(b->class));
-            log_tag("debug_log.txt", "[FIGHT-BOSS]", msg);
-        }
-    } else {
-        //Apply status effects to boss
-        if (b->status != Normal) {
-            applyBStatus(notify_win, b);
-            sprintf(msg, "Applied  %s to %s.", stringFromStatus(b->status),
-                    stringFromBossClass(b->class));
-            log_tag("debug_log.txt", "[STATUS]", msg);
-        }
-    }
-
-    if (player->hp <= 0) {
-        res = FIGHTRES_DEATH;
-    } else {
-        //Apply status effects to player
-        if (player->status != Normal) {
-            applyStatus(notify_win, player);
-        }
-    }
-
-    //Consumable drop, guaranteed on killing a beast
-    if (res == FIGHTRES_KILL_DONE) {
-        int drop = dropConsumable(player);
-        wattron(notify_win, COLOR_PAIR(S4C_CYAN));
-        sprintf(msg, "You found a %s!", stringFromConsumables(drop));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_CYAN));
-        sprintf(msg, "Found Consumable:    %s.", stringFromConsumables(drop));
-        log_tag("debug_log.txt", "[DROPS]", msg);
-    }
-
-    //Artifact drop (if we don't have all of them)
-    if (res == FIGHTRES_KILL_DONE
-        && (player->stats->artifactsfound != ARTIFACTSMAX + 1)) {
-        int artifact_drop = dropArtifact(player);
-        wattron(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        sprintf(msg, "You found a %s!", stringFromArtifacts(artifact_drop));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        sprintf(msg, "Found Artifact:    %s.",
-                stringFromArtifacts(artifact_drop));
-        log_tag("debug_log.txt", "[DROPS]", msg);
-    }
-    //Equip drop
-    if (res == FIGHTRES_KILL_DONE) {
-        //We give 1 to obtain the better equip generation used for beasts
-        dropEquip(player, 1, notify_win, kls);
-    }
-
-    return res;
-}
-
-int boss_attack(Boss *b, Fighter *target, Path *p, WINDOW *notify_win,
-                Koliseo *kls)
-{
-
-    //TODO
-    //Implementation similar to boss_fight(), as a base idea
-    //Should return fightResult values, while keeping the perspective on the Fighter, as in:
-    //
-    //      FIGHTRES_DEATH means the Fighter died
-    //      FIGHTRES_KILL_DONE means the Boss died
-    //      FIGHTRES_DMG_DEALT means the Fighter inflicted damage
-    //      FIGHRES_DMG_TAKEN means the Fighter received damage
-    fightResult res = FIGHTRES_NO_DMG;
-    //Stat comparisons
-    //
-    char msg[200];
-
-    int atkdelta = (b->atk + b->turnboost_atk) - (target->atk + target->equipboost_atk + target->turnboost_atk - (rand() % 3)) - 1;	//Skewed with defender
-    int defdelta = (b->def + b->turnboost_def) - (target->def + target->equipboost_def + target->turnboost_def) + (rand() % 2) + 1;	//Skewed with attacker
-    int veldelta =
-        (b->vel + b->turnboost_vel) - (target->vel + target->equipboost_vel +
-                                       target->turnboost_vel) + (rand() % 3) +
-        1;
-
-    int atkOnPlayer =
-        (b->atk + b->turnboost_atk) - (target->def + target->equipboost_def +
-                                       target->turnboost_def +
-                                       (target->vel / 6));
-    int atkOnEnemy =
-        (target->atk + target->equipboost_atk + target->turnboost_atk) -
-        (b->def + b->turnboost_def + (b->vel / 6));
-    if (G_GODMODE_ON == 1) {
-        log_tag("debug_log.txt", "[DEBUG]", "[%s]:  G_GODMODE_ON == 1",
-                __func__);
-        atkdelta = -100;
-        defdelta = -100;
-        veldelta = -100;
-        atkOnPlayer = 1;
-        atkOnEnemy = 100;
-    }
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "atkdelta %i", atkdelta);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "defdelta %i", defdelta);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "veldelta %i", veldelta);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "atkOnEnemy %i", atkOnEnemy);
-    log_tag("debug_log.txt", "[DEBUG-FIGHT]", "atkOnPlayer %i", atkOnPlayer);
-
-    int damageDealt = -1;
-
-    if (veldelta >= 0) {	//Boss was faster
-        if (atkOnPlayer > 3) {
-            damageDealt = atkOnPlayer;
-            target->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result D LOST (slower, great enemy atk.");
-        } else if (atkOnPlayer >= 0) {
-            damageDealt = abs(atkOnPlayer - atkdelta);
-            target->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result E LOST (slower, ok enemy atk.");
-        } else {
-            if (atkOnPlayer < -3) {
-                damageDealt = fabsf(atkOnEnemy - 0.75F * b->vel);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result F1 WIN (slower, enemy atk < -3.");
-            } else {
-                damageDealt = abs(atkOnEnemy - 1);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result F2 WIN (slower, enemy atk > -3.");
-            }
-            b->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-        }
-    } else {			//Target was faster
-        atkdelta = -atkdelta;
-        if (atkOnEnemy > 3) {
-            damageDealt = atkOnEnemy;
-            b->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result A WIN (faster, great atk).");
-        } else if (atkOnEnemy >= 0) {
-            damageDealt = abs(atkOnEnemy - atkdelta);
-            b->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_DEALT;
-            log_tag("debug_log.txt", "[FIGHT]",
-                    "Fight result B WIN (faster, ok atk).");
-        } else {
-            if (atkOnEnemy < -3) {
-                damageDealt =
-                    fabsf(atkOnPlayer -
-                          0.75F * (target->vel + target->equipboost_vel));
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result C1 LOST (faster, atk < -3).");
-            } else {
-                damageDealt = abs(atkOnPlayer - 1);
-                log_tag("debug_log.txt", "[FIGHT]",
-                        "Fight result C2 LOST (faster, atk > -3).");
-            }
-            target->hp -= damageDealt > 0 ? damageDealt : 1;
-            res = FIGHTRES_DMG_TAKEN;
-        }
-    }
-
-    log_tag("debug_log.txt", "[FIGHT]", "damageCalc %i", damageDealt);
-
-    int yourhit = (res == FIGHTRES_DMG_DEALT) ? 1 : 0;
-    char victim[25];
-
-    if (!yourhit) {
-
-        b->vel--;
-        b->atk--;
-        b->def -= 2;
-
-        //Check if someone earned a stat reset after the fight
-        statReset(target, 0);
-        statResetBoss(b, 0);
-
-        strcpy(victim, target->name);
-    } else {
-
-        target->vel--;
-        target->atk--;
-        target->def -= 2;
-
-        //Account for vampirism perk
-        int vampire_perks = target->perks[VAMPIRISM]->innerValue;
-        if (vampire_perks > 0) {
-            int recovery = floor(damageDealt * (0.1 * vampire_perks));
-            target->hp += recovery;
-            log_tag("debug_log.txt", "[PERKS]", "Vampirism proc for +%i HP.",
-                    recovery);
-            if (target->hp >= target->totalhp) {
-                target->hp = target->totalhp;
-            };
-        }
-        //Account for burn on touch perk
-        int hotbody_perks = target->perks[HOT_BODY]->innerValue;
-        if (hotbody_perks > 0) {
-            int burnchance = 11 - hotbody_perks;
-            if (rand() % burnchance == 0) {
-                b->status = Burned;	//Set status to Burned. May need change to manage multiple statuses active at once
-                setCounter((Turncounter *) b->counters[Burned], 2);	//Give 2 turns of Burned status
-                log_tag("debug_log.txt", "[PERKS]",
-                        "Hotbody proc on 1/%i chance.", burnchance);
-            }
-        }
-        //Account for poison on touch perk. Order of checks with hot_body perk may cause issues?
-        int biohazard_perks = target->perks[BIOHAZARD]->innerValue;
-        if (biohazard_perks > 0) {
-            int poisonchance = 11 - biohazard_perks;
-            if (rand() % poisonchance == 0) {
-                b->status = Poison;	//Set status to Poison. May need change to manage multiple statuses active at once
-                setCounter((Turncounter *) b->counters[POISON], 2);	//Give 2 turns of Poison status
-                log_tag("debug_log.txt", "[PERKS]",
-                        "Biohazard proc on 1/%i chance.", poisonchance);
-            }
-        }
-        //Check if someone earned a stat reset after the fight
-        statResetBoss(b, 0);
-        statReset(target, 0);
-
-        strcpy(victim, stringFromBossClass(b->class));
-    }
-
-    int color = -1;
-    if (yourhit) {
-        color = S4C_WHITE;
-    } else {
-        color = S4C_RED;
-    }
-
-    wattron(notify_win, COLOR_PAIR(color));
-    sprintf(msg, "%s was hit.    (%i DMG)", victim,
-            damageDealt > 0 ? damageDealt : 1);
-    display_notification(notify_win, msg, 500);
-    wattroff(notify_win, COLOR_PAIR(color));
-
-    //Rolls
-    //
-    //Critical hit roll
-
-    //Account for critboost_chance perks
-    int critMax =
-        round(10.0 - floor(target->luck / 5) -
-              (1.5 * target->perks[CRITBOOST_CHANCE]->innerValue));
-
-    int critRes = (rand() % critMax);
-
-    if (res == FIGHTRES_DMG_DEALT && (critRes <= 0)) {
-
-        //Account for critboost_dmg perks
-        int dmgboost_perks = target->perks[CRITBOOST_DMG]->innerValue;
-        damageDealt *= (0.30 + (0.12 * dmgboost_perks));
-        b->hp -= damageDealt > 0 ? damageDealt : 1;
-        wattron(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        sprintf(msg, "A critical hit!    (%i DMG)",
-                damageDealt > 0 ? damageDealt : 1);
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        log_tag("debug_log.txt", "[FIGHT-BOSS]",
-                "Critical hit for %i dmg, proc on 1/%i chance.", damageDealt,
-                critMax);
-
-        //Update stats
-        target->stats->criticalhits++;
-    }
-    //Check for deaths -> exit condition from loop
-    //
-    //
-    //
-    if (b->hp <= 0) {
-        res = FIGHTRES_KILL_DONE;
-
-        //Account for runic circle perk
-        int runic_perks = target->perks[RUNIC_MAGNET]->innerValue;
-        if (runic_perks > 0) {
-            int recovery = round(0.51 * runic_perks);
-            target->energy += recovery;
-            log_tag("debug_log.txt", "[PERKS]",
-                    "Runicmagnet proc for %i energy.", recovery);
-        }
-        if (b->beast) {
-            color = S4C_MAGENTA;
-        } else {
-            color = S4C_RED;
-        }
-        wattron(notify_win, COLOR_PAIR(color));
-        sprintf(msg, "%s fainted.", stringFromBossClass(b->class));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(color));
-        log_tag("debug_log.txt", "[FIGHT-BOSS]", "Killed  %s.",
-                stringFromBossClass(b->class));
-
-        //Update stats
-        target->stats->bosseskilled++;
-
-        //Check if we killed this boss before
-        if (target->stats->killed_bosses[b->class] == 0) {
-            //Update stats
-            target->stats->unique_bosseskilled++;
-            target->stats->killed_bosses[b->class] += 1;
-            //Check if we have to update the wincon value
-            if (p->win_condition->class == ALL_BOSSES) {
-                p->win_condition->current_val++;
-            }
-            log_tag("debug_log.txt", "[FIGHT-BOSS]", "Killed new boss  %s.",
-                    stringFromBossClass(b->class));
-        }
-    } else {
-        //Apply status effects to boss
-        if (b->status != Normal) {
-            applyBStatus(notify_win, b);
-            log_tag("debug_log.txt", "[STATUS]", "Applied  %s to %s.",
-                    stringFromStatus(b->status), stringFromBossClass(b->class));
-        }
-    }
-
-    if (target->hp <= 0) {
-        res = FIGHTRES_DEATH;
-    } else {
-        //Apply status effects to target
-        if (target->status != Normal) {
-            applyStatus(notify_win, target);
-        }
-    }
-
-    //Consumable drop, guaranteed on killing a beast
-    if (res == FIGHTRES_KILL_DONE) {
-        int drop = dropConsumable(target);
-        wattron(notify_win, COLOR_PAIR(S4C_CYAN));
-        sprintf(msg, "You found a %s!", stringFromConsumables(drop));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_CYAN));
-        log_tag("debug_log.txt", "[DROPS]", "Found Consumable:    %s.",
-                stringFromConsumables(drop));
-    }
-
-    //Artifact drop (if we don't have all of them)
-    if (res == FIGHTRES_KILL_DONE
-        && (target->stats->artifactsfound != ARTIFACTSMAX + 1)) {
-        int artifact_drop = dropArtifact(target);
-        wattron(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        sprintf(msg, "You found a %s!", stringFromArtifacts(artifact_drop));
-        display_notification(notify_win, msg, 500);
-        wattroff(notify_win, COLOR_PAIR(S4C_MAGENTA));
-        log_tag("debug_log.txt", "[DROPS]", "Found Artifact:    %s.",
-                stringFromArtifacts(artifact_drop));
-    }
-    //Equip drop
-    if (res == FIGHTRES_KILL_DONE) {
-        //We give 1 to obtain the better equip generation used for beasts
-        dropEquip(target, 1, notify_win, kls);
-    }
-
-    return res;
-}
-
-/**
- * Takes a Fighter and a Boss pointers and calls boss_fight().
- * @see Fighter
- * @see Boss
- * @see boss_fight()
- * @param player The Fighter pointer at hand.
- * @param b The Boss pointer at hand.
- * @param foe_op The foeTurnOption_OP for the foe.
- * @param notify_win The WINDOW pointer to call display_notification() on.
- * @param kls The Koliseo used for allocations.
- */
-int defer_fight_boss(Fighter *player, Boss *b, Path *p, foeTurnOption_OP foe_op,
-                     WINDOW *notify_win, Koliseo *kls)
-{
-    char msg[200];
-    //FIXME
-    //Is it okay to return just one result, when having 2 interactions that could go differently?
-    //
-    //Use FIGHTRES_CLASH as needed, to indicate both sides were damaged at some point.
-    fightResult res = FIGHTRES_NO_DMG;
-
-    int player_goes_first = (player->vel >= b->vel ? 1 : 0);
-
-    if (player_goes_first) {
-        res = boss_fight(player, b, p, notify_win, kls);
-        //Check res and apply second action if needed
-        log_tag("debug_log.txt", "[DEBUG]",
-                "[%s()]: First act res was [%s]: [%i]", __func__,
-                stringFrom_fightResult(res), res);
-        if (res != FIGHTRES_DEATH && res != FIGHTRES_KILL_DONE) {
-            switch (foe_op) {
-            case FOE_OP_INVALID: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "foe_op was FOE_OP_INVALID in [%s]: [%i]", __func__,
-                        foe_op);
-                kls_free(default_kls);
-                kls_free(temporary_kls);
-                exit(EXIT_FAILURE);
-            }
-            break;
-            case FOE_OP_IDLE: {
-                log_tag("debug_log.txt", "[DEFER]",
-                        "[%s()]:  Foe { %s } was idle.", __func__,
-                        stringFromBossClass(b->class));
-                wattron(notify_win, COLOR_PAIR(S4C_GREY));
-                sprintf(msg, "%s is loafing around.",
-                        stringFromBossClass(b->class));
-                display_notification(notify_win, msg, 500);
-                wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-            }
-            break;
-            case FOE_OP_FIGHT: {
-                log_tag("debug_log.txt", "[DEFER]",
-                        "[%s()]:  Foe { %s } wants to fight.", __func__,
-                        stringFromBossClass(b->class));
-                wattron(notify_win, COLOR_PAIR(S4C_GREY));
-                sprintf(msg, "%s is angry!", stringFromBossClass(b->class));
-                display_notification(notify_win, msg, 500);
-                wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-                res = boss_attack(b, player, p, notify_win, kls);
-            }
-            break;
-            case FOE_OP_SPECIAL: {
-                log_tag("debug_log.txt", "[TODO]",
-                        "[%s()]:  Foe { %s } wants to use a special.",
-                        __func__, stringFromBossClass(b->class));
-                //TODO
-                //Implement boss special function
-                //res = boss_attack_special(b,player,p,notify_win,kls);
-            }
-            break;
-            default: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "Unexpected foeTurnOption_OP in [%s()]: [%i]",
-                        __func__, foe_op);
-                kls_free(default_kls);
-                kls_free(temporary_kls);
-                exit(EXIT_FAILURE);
-            }
-            break;
-            }			// End foe_op switch
-
-            //TODO
-            //Check second turn act res?
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "[%s()]: Second act res was [%s]: [%i]", __func__,
-                    stringFrom_fightResult(res), res);
-
-            return res;
-        } else if (res == FIGHTRES_DEATH) {
-            return res;
-        } else if (res == FIGHTRES_KILL_DONE) {
-            return res;
-        }
-    } else {
-        //Foe acts first
-        switch (foe_op) {
-        case FOE_OP_INVALID: {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "foe_op was FOE_OP_INVALID in [%s()]: [%i]", __func__,
-                    foe_op);
-            kls_free(default_kls);
-            kls_free(temporary_kls);
-            exit(EXIT_FAILURE);
-        }
-        break;
-        case FOE_OP_IDLE: {
-            log_tag("debug_log.txt", "[DEFER]",
-                    "[%s()]:  Foe { %s } was idle.", __func__,
-                    stringFromBossClass(b->class));
-            wattron(notify_win, COLOR_PAIR(S4C_GREY));
-            sprintf(msg, "%s is loafing around.",
-                    stringFromBossClass(b->class));
-            display_notification(notify_win, msg, 500);
-            wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-        }
-        break;
-        case FOE_OP_FIGHT: {
-            log_tag("debug_log.txt", "[DEFER]",
-                    "[%s()]:  Foe { %s } wants to fight.", __func__,
-                    stringFromBossClass(b->class));
-            wattron(notify_win, COLOR_PAIR(S4C_GREY));
-            sprintf(msg, "%s is angry!", stringFromBossClass(b->class));
-            display_notification(notify_win, msg, 500);
-            wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-            res = boss_attack(b, player, p, notify_win, kls);
-        }
-        break;
-        case FOE_OP_SPECIAL: {
-            log_tag("debug_log.txt", "[TODO]",
-                    "[%s()]:  Foe { %s } wants to use a special.", __func__,
-                    stringFromBossClass(b->class));
-            //TODO
-            //Implement boss special function
-            //res = boss_attack_special(b,player,p,notify_win,kls);
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Unexpected foeTurnOption_OP in [%s()]: [%i]", __func__,
-                    foe_op);
-            kls_free(default_kls);
-            kls_free(temporary_kls);
-            exit(EXIT_FAILURE);
-        }
-        break;
-        }			// End foe_op switch
-
-        //Check res and apply second action if needed
-        log_tag("debug_log.txt", "[DEBUG]",
-                "[%s()]: First act res was [%s]: [%i]", __func__,
-                stringFrom_fightResult(res), res);
-
-        if (res != FIGHTRES_DEATH && res != FIGHTRES_KILL_DONE) {
-            res = boss_fight(player, b, p, notify_win, kls);
-
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "[%s()]: Second act res was [%s]: [%i]", __func__,
-                    stringFrom_fightResult(res), res);
-            //TODO
-            //Check second turn act res?
-            return res;
-
-        } else if (res == FIGHTRES_DEATH) {
-            return res;
-        } else if (res == FIGHTRES_KILL_DONE) {
-            return res;
-        }
-    }
-
-    return res;
-}
-
-/**
- * Takes a Fighter and a Boss pointers and calls do_Skill_boss().
- * @see Fighter
- * @see Boss
- * @see do_Skill_boss()
- * @param player The Fighter pointer at hand.
- * @param b The Boss pointer at hand.
- * @param picked_skill The skill picked by Fighter.
- * @param foe_op The foeTurnOption_OP for the foe.
- * @param notify_win The WINDOW pointer to call display_notification() on.
- * @param kls The Koliseo used for allocations.
- */
-int defer_skill_boss(Fighter *player, Boss *b, skillType picked_skill, Path *p, foeTurnOption_OP foe_op,
-                     WINDOW *notify_win, Koliseo *kls)
-{
-    char msg[200];
-    //FIXME
-    //Is it okay to return just one result, when having 2 interactions that could go differently?
-    //
-    //Use FIGHTRES_CLASH as needed, to indicate both sides were damaged at some point.
-    fightResult res = FIGHTRES_NO_DMG;
-
-    int player_goes_first = (player->vel >= b->vel ? 1 : 0);
-
-    if (player_goes_first) {
-        res = do_Skill_boss(player, b, picked_skill, p, notify_win, kls);
-        //Check res and apply second action if needed
-        log_tag("debug_log.txt", "[DEBUG]",
-                "[%s()]: First act res was [%s]: [%i]", __func__,
-                stringFrom_fightResult(res), res);
-        if (res != FIGHTRES_DEATH && res != FIGHTRES_KILL_DONE) {
-            switch (foe_op) {
-            case FOE_OP_INVALID: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "foe_op was FOE_OP_INVALID in [%s]: [%i]", __func__,
-                        foe_op);
-                kls_free(default_kls);
-                kls_free(temporary_kls);
-                exit(EXIT_FAILURE);
-            }
-            break;
-            case FOE_OP_IDLE: {
-                log_tag("debug_log.txt", "[DEFER]",
-                        "[%s()]:  Foe { %s } was idle.", __func__,
-                        stringFromBossClass(b->class));
-                wattron(notify_win, COLOR_PAIR(S4C_GREY));
-                sprintf(msg, "%s is loafing around.",
-                        stringFromBossClass(b->class));
-                display_notification(notify_win, msg, 500);
-                wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-            }
-            break;
-            case FOE_OP_FIGHT: {
-                log_tag("debug_log.txt", "[DEFER]",
-                        "[%s()]:  Foe { %s } wants to fight.", __func__,
-                        stringFromBossClass(b->class));
-                wattron(notify_win, COLOR_PAIR(S4C_GREY));
-                sprintf(msg, "%s is angry!", stringFromBossClass(b->class));
-                display_notification(notify_win, msg, 500);
-                wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-                res = boss_attack(b, player, p, notify_win, kls);
-            }
-            break;
-            case FOE_OP_SPECIAL: {
-                log_tag("debug_log.txt", "[TODO]",
-                        "[%s()]:  Foe { %s } wants to use a special.",
-                        __func__, stringFromBossClass(b->class));
-                //TODO
-                //Implement boss special function
-                //res = boss_attack_special(b,player,p,notify_win,kls);
-            }
-            break;
-            default: {
-                log_tag("debug_log.txt", "[ERROR]",
-                        "Unexpected foeTurnOption_OP in [%s()]: [%i]",
-                        __func__, foe_op);
-                kls_free(default_kls);
-                kls_free(temporary_kls);
-                exit(EXIT_FAILURE);
-            }
-            break;
-            }			// End foe_op switch
-
-            //TODO
-            //Check second turn act res?
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "[%s()]: Second act res was [%s]: [%i]", __func__,
-                    stringFrom_fightResult(res), res);
-
-            return res;
-        } else if (res == FIGHTRES_DEATH) {
-            return res;
-        } else if (res == FIGHTRES_KILL_DONE) {
-            return res;
-        }
-    } else {
-        //Foe acts first
-        switch (foe_op) {
-        case FOE_OP_INVALID: {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "foe_op was FOE_OP_INVALID in [%s()]: [%i]", __func__,
-                    foe_op);
-            kls_free(default_kls);
-            kls_free(temporary_kls);
-            exit(EXIT_FAILURE);
-        }
-        break;
-        case FOE_OP_IDLE: {
-            log_tag("debug_log.txt", "[DEFER]",
-                    "[%s()]:  Foe { %s } was idle.", __func__,
-                    stringFromBossClass(b->class));
-            wattron(notify_win, COLOR_PAIR(S4C_GREY));
-            sprintf(msg, "%s is loafing around.",
-                    stringFromBossClass(b->class));
-            display_notification(notify_win, msg, 500);
-            wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-        }
-        break;
-        case FOE_OP_FIGHT: {
-            log_tag("debug_log.txt", "[DEFER]",
-                    "[%s()]:  Foe { %s } wants to fight.", __func__,
-                    stringFromBossClass(b->class));
-            wattron(notify_win, COLOR_PAIR(S4C_GREY));
-            sprintf(msg, "%s is angry!", stringFromBossClass(b->class));
-            display_notification(notify_win, msg, 500);
-            wattroff(notify_win, COLOR_PAIR(S4C_GREY));
-            res = boss_attack(b, player, p, notify_win, kls);
-        }
-        break;
-        case FOE_OP_SPECIAL: {
-            log_tag("debug_log.txt", "[TODO]",
-                    "[%s()]:  Foe { %s } wants to use a special.", __func__,
-                    stringFromBossClass(b->class));
-            //TODO
-            //Implement boss special function
-            //res = boss_attack_special(b,player,p,notify_win,kls);
-        }
-        break;
-        default: {
-            log_tag("debug_log.txt", "[ERROR]",
-                    "Unexpected foeTurnOption_OP in [%s()]: [%i]", __func__,
-                    foe_op);
-            kls_free(default_kls);
-            kls_free(temporary_kls);
-            exit(EXIT_FAILURE);
-        }
-        break;
-        }			// End foe_op switch
-
-        //Check res and apply second action if needed
-        log_tag("debug_log.txt", "[DEBUG]",
-                "[%s()]: First act res was [%s]: [%i]", __func__,
-                stringFrom_fightResult(res), res);
-
-        if (res != FIGHTRES_DEATH && res != FIGHTRES_KILL_DONE) {
-            res = do_Skill_boss(player, b, picked_skill, p, notify_win, kls);
-
-            log_tag("debug_log.txt", "[DEBUG]",
-                    "[%s()]: Second act res was [%s]: [%i]", __func__,
-                    stringFrom_fightResult(res), res);
-            //TODO
-            //Check second turn act res?
-            return res;
-
-        } else if (res == FIGHTRES_DEATH) {
-            return res;
-        } else if (res == FIGHTRES_KILL_DONE) {
-            return res;
-        }
-    }
-
-    return res;
-}
-/**
  * Takes a Fighter, an Enemy and a Boss pointers, a string denoting the consumableClass and an int for use on enemy or boss.
  * If qty value for the Consumable is 0, we have an early return. Otherise effect is applied and qty is decreased.
  * @see Fighter
@@ -3365,1393 +478,6 @@ int retry(void)
         kls_free(temporary_kls);
         exit(EXIT_FAILURE);
     }
-}
-
-/**
- * Takes a Fighter and a Path pointers (and an integer for current room index) and asks user input to execute debug actions.
- * @see Fighter
- * @see Path
- * @see statReset()
- * @see GET_CALLBACK()
- * @see unlock_special()
- * @param gmst The Gamestate pointer.
- * @param player The Fighter pointer at hand.
- * @param p The Path pointer of the current game.
- * @param roomIndex The index of current room.
- * @param The Koliseo to debug.
- * @param The Koliseo_Temp used for allocations.
- */
-void debug_generic(Gamestate *gmst, Fighter *player, Path *p, int roomIndex,
-                   Koliseo *kls, Koliseo_Temp *t_kls)
-{
-
-    char msg[200];
-    char ch[25];
-    int picked_debug_proc = 0;
-#ifndef _WIN32
-    struct utsname uts;
-    uname(&uts);
-    sprintf(msg, "debug_generic() loaded utsname using uname().\n");
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-    sprintf(msg, "System is %s\n", uts.sysname);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-    sprintf(msg, "OS Release is %s\n", uts.release);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-    sprintf(msg, "OS Version is %s\n", uts.version);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-    sprintf(msg, "Machine is %s\n", uts.machine);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-#endif
-
-    int res = system("clear");
-    sprintf(msg, "debug_generic() system(\"clear\") res was (%i)", res);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-
-    int c = 0, n = -1;
-    while (!picked_debug_proc) {
-        int res = system("clear");
-        sprintf(msg, "debug_generic() 2 system(\"clear\") res was (%i)", res);
-        log_tag("debug_log.txt", "[DEBUG]", msg);
-        printf("\n\
-	What do you want to do? ('q' to quit)\n\
-		[\n\
-			'0':  Give xp points\t'1':  Give consumable\n\
-			'2':  Reset stats\t'3':  Alter luck\n\
-			'4':  Give coins\t'5':  Unlock special\n\
-			'6':  Unlock Artifact\t'7':   Print turnCounter\n\
-			's':  Sprites slideshow\t'd':  Dump debug symbols\n\
-			'g':  Toggle godmode\t'A':  Toggle autosave\n\
-			'L':  Toggle logging\t'F':  Try Floor functionality\n\
-			'Q':  Toggle fast quit\t'k': Log passed kls state to debug log file.\n\
-			't': Log global temporary_kls state to debug log file.\n\
-			'K': Log usage stats for passed kls to debug log file.\n\
-			'T': Log usage stats for temporary_kls to debug log file.\n\
-			'G': Log Gamestate to debug log file.\n\
-			{Return}  Process your input line.\t'q':  Quit\n\
-		]\n\n\
-	[%s@debug-func]$ ", player->name);
-
-        char *fgetsres = fgets(ch, sizeof ch, stdin);
-        sprintf(msg, "debug_generic() fgets() res was (%s)", fgetsres);
-        log_tag("debug_log.txt", "[DEBUG]", msg);
-        switch (ch[0]) {
-        case '0': {
-            picked_debug_proc = 1;
-            int xp;
-            do {
-                printf("\nHow much?\n");
-            } while ((c = scanf("%i", &xp)) != 1 || xp < 0);
-            int res = scanf("%*c");
-            sprintf(msg, "debug_generic() scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            checkremainder(player, xp);
-        }
-        break;
-        case '1': {
-            picked_debug_proc = 1;
-            int q = -1;
-            do {
-                printf("\nInsert consumable number:\n");
-            } while ((c = scanf("%i", &n)) != 1 || n > CONSUMABLESMAX
-                     || n < 0);
-            int res = scanf("%*c");
-            sprintf(msg, "debug_generic() scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            do {
-                printf("\nInsert quantity:\n");
-            } while ((c = scanf("%i", &q)) != 1 && q <= 0);
-            res = scanf("%*c");
-            sprintf(msg, "debug_generic() 2 scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            Consumable *c = (Consumable *) player->consumablesBag[n];
-            c->qty += q;
-            player->stats->consumablesfound += q;
-        }
-        break;
-        case '2': {
-            picked_debug_proc = 1;
-            statReset(player, 1);
-        }
-        break;
-        case '3': {
-            picked_debug_proc = 1;
-            printf("\nCurrent luck: %i\tRL: %i\n", player->luck, p->luck);
-
-            do {
-                printf("\nInsert new luck:\n");
-            } while ((c = scanf("%i", &n)) != 1 && n > 0);
-            int res = scanf("%*c");
-            sprintf(msg, "debug_generic() 3 scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            //FIXME:
-            //Copy-pasted the calc for player luck... might need a function
-            p->luck = n;
-
-            player->luck = (p->luck * MAXPLAYERLUCK) / MAXLUCK;
-        }
-        break;
-        case '4': {
-            picked_debug_proc = 1;
-            /* Old code invoking the macro for special moves directly. Used for testing.
-             * int s = -1;
-             * printf("Insert special num:\n");
-             * scanf("%i",&s);
-             * scanf("%*c");
-             * printf("Read: %i\n",s);
-             * GET_CALLBACK(s,callback_special_t)(player,e,p,roomIndex,currentEnemyNum);
-             */
-
-            //printActivePerks(player);
-            int c;
-            int n = -1;
-            do {
-                printf("\nInsert coin number (0<n<100):\n");
-            } while ((c = scanf("%i", &n)) < 0 || c > 3 || n < 0
-                     || n > 100);
-            int res = scanf("%*c");
-            sprintf(msg, "debug_generic() 4 scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            player->balance += n;
-        }
-        break;
-        case '5': {
-            picked_debug_proc = 1;
-            unlockSpecial(player);
-        }
-        break;
-        case '6': {
-            picked_debug_proc = 1;
-            int n = -1;
-            do {
-                printf("\nInsert artifact number (0<n<%i):\n",
-                       ARTIFACTSMAX + 1);
-            } while ((c = scanf("%i", &n)) != 1 || n > ARTIFACTSMAX
-                     || n < 0);
-
-            int res = scanf("%*c");
-            sprintf(msg, "debug_generic() scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            player->artifactsBag[n]->qty += 1;
-            player->artifactsBag[n]->active = 0;
-            player->stats->artifactsfound += 1;
-        }
-        break;
-        case '7': {
-            picked_debug_proc = 1;
-            int res = system("clear");
-            sprintf(msg, "debug_generic() 3 system(\"clear\") res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            printf("\nPlayer Counters:\n");
-            printCounters((Turncounter **) player->counters);
-            printf("\nPress Enter to resume game");
-            res = scanf("%*c");
-            sprintf(msg, "debug_generic() 5 scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 'd': {
-            picked_debug_proc = 1;
-            printf("\nVERSION:    %s\n", VERSION);
-#ifndef _WIN32
-            printf("\nSystem:    %s\n", uts.sysname);
-            printf("\nOS Release:    %s\n", uts.release);
-            printf("\nOS Version:    %s\n", uts.version);
-            printf("\nMachine:    %s\n", uts.machine);
-#endif
-            printf("\nGAMEMODE:    %s\n", stringFromGamemode(GAMEMODE));
-            printf("\nPath->current_saveslot->save_path:    %s\n",
-                   p->current_saveslot->save_path);
-            printf("\nGS_AUTOSAVE_ON:    %i\n", GS_AUTOSAVE_ON);
-            printf("\nG_FASTQUIT_ON:    %i\n", G_FASTQUIT_ON);
-            printf("\nG_DEBUG_ON:    %i\n", G_DEBUG_ON);
-            printf("\nG_LOG_ON:    %i\n", G_LOG_ON);
-            printf("\nG_GODMODE_ON:    %i\n", G_GODMODE_ON);
-            printf("\nG_EXPERIMENTAL_ON:    %i\n", G_EXPERIMENTAL_ON);
-            printf("\nG_DEBUG_ROOMTYPE_ON:    %i\n", G_DEBUG_ROOMTYPE_ON);
-            if (G_DEBUG_ROOMTYPE_ON == 1) {
-                printf("\nG_DEBUG_ROOMTYPE:    %i\n", G_DEBUG_ROOMTYPE);
-            }
-            printf("\nG_DEBUG_ENEMYTYPE_ON:    %i\n", G_DEBUG_ENEMYTYPE_ON);
-            if (G_DEBUG_ENEMYTYPE_ON == 1) {
-                printf("\nG_DEBUG_ENEMYTYPE:    %i\n", G_DEBUG_ENEMYTYPE);
-            }
-            printf("\nPress Enter to resume game.\n");
-            int res = scanf("%*c");
-            sprintf(msg, "debug_generic() 7 scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 'F': {
-            sprintf(msg, "Trying out Floor functionality.");
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            // Declare dbg_floor
-            sprintf(msg, "Pushing dbg_floor to tkls.");
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            kls_log(kls, "DEBUG", msg);
-            Floor *dbg_floor =
-                (Floor *) KLS_PUSH_T_TYPED(t_kls, Floor, HR_Floor,
-                                           "Floor", msg);
-            // Start the random walk from the center of the dungeon
-            int center_x = FLOOR_MAX_COLS / 2;
-            int center_y = FLOOR_MAX_ROWS / 2;
-            // Init dbg_floor
-            init_floor_layout(dbg_floor);
-            //Set center as filled
-            dbg_floor->floor_layout[center_x][center_y] = 1;
-
-            init_floor_rooms(dbg_floor);
-
-            floor_random_walk(dbg_floor, center_x, center_y, 100, 1);	// Perform 100 steps of random walk, reload floor_layout if needed
-            load_floor_explored(dbg_floor);
-
-            floor_set_room_types(dbg_floor);
-
-            /*
-               debug_print_floor_layout(dbg_floor);
-
-               printf("\nPress Enter to see room layout.\n");
-               int res = scanf("%*c");
-               sprintf(msg,"debug_generic() 7 scanf() res was (%i)", res);
-               log_tag("debug_log.txt","[DEBUG]",msg);
-             */
-            debug_print_roomclass_layout(dbg_floor, stdout);
-            printf("\nPress Enter to return to game.\n");
-            res = scanf("%*c");
-            sprintf(msg, "debug_generic() 8 scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            //free(dbg_floor);
-        }
-        break;
-        case 'A': {
-            GS_AUTOSAVE_ON = (GS_AUTOSAVE_ON == 1 ? 0 : 1);
-            sprintf(msg, "Toggled G_AUTOSAVE_ON, new value: (%i)",
-                    GS_AUTOSAVE_ON);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 'G': {
-            log_tag("debug_log.txt", "[DEBUG]", "Logging Gamestate");
-            dbg_Gamestate(gmst);
-            log_tag("debug_log.txt", "[DEBUG]", "Done logging Gamestate");
-        }
-        break;
-        case 'T': {
-            char path_to_kls_file[600];
-            char static_path[500];
-            // Set static_path value to the correct static dir path
-            resolve_staticPath(static_path);
-
-            //Append to "kls_log.txt"
-            sprintf(path_to_kls_file, "%s/%s", static_path,
-                    "debug_log.txt");
-            FILE *kls_file = NULL;
-            kls_file = fopen(path_to_kls_file, "a");
-            if (kls_file == NULL) {
-                fprintf(stderr,
-                        "debug_generic():  failed opening debug logfile.\n");
-                exit(EXIT_FAILURE);
-            }
-            if (kls == NULL) {
-                fprintf(stderr, "debug_generic():  kls was NULL.\n");
-                exit(EXIT_FAILURE);
-            }
-
-            fprintf(kls_file, "--BEGIN debug of temporary_kls--\n");
-            for (int i = HR_Path; i < HLP_MAX_INDEX + 99; i++) {
-                ptrdiff_t usage = kls_type_usage(i, temporary_kls);
-#ifndef _WIN32
-                fprintf(kls_file,
-                        "Usage for HLP_Region_Type { %s } [Index: %i]  {Size: %li }\n",
-                        stringFrom_HLP_Region_Type(i - 101 +
-                                                   KLS_REGIONTYPE_MAX), i,
-                        usage);
-#else
-                fprintf(kls_file,
-                        "Usage for HLP_Region_Type { %s } [Index: %i]  {Size: %lli }\n",
-                        stringFrom_HLP_Region_Type(i - 101 +
-                                                   KLS_REGIONTYPE_MAX), i,
-                        usage);
-#endif
-            }
-            fprintf(kls_file, "--END debug of temporary_kls--\n\n");
-
-            fclose(kls_file);
-        }
-        break;
-        case 't': {
-            char path_to_kls_file[600];
-            char static_path[500];
-            // Set static_path value to the correct static dir path
-            resolve_staticPath(static_path);
-
-            //Append to "kls_log.txt"
-            sprintf(path_to_kls_file, "%s/%s", static_path,
-                    "debug_log.txt");
-            FILE *kls_file = NULL;
-            kls_file = fopen(path_to_kls_file, "a");
-            if (kls_file == NULL) {
-                fprintf(stderr,
-                        "debug_generic():  failed opening debug logfile.\n");
-                exit(EXIT_FAILURE);
-            }
-            if (kls == NULL) {
-                fprintf(stderr, "debug_generic():  kls was NULL.\n");
-                exit(EXIT_FAILURE);
-            }
-            fprintf(kls_file, "--BEGIN debug of temporary_kls--\n");
-            print_kls_2file(kls_file, temporary_kls);
-            kls_showList_toFile(temporary_kls->regs, kls_file);
-            kls_usageReport_toFile(temporary_kls, kls_file);
-            fprintf(kls_file, "--END debug of temporary_kls--\n\n");
-            WINDOW *win = NULL;
-            /* Initialize curses */
-            clear();
-            refresh();
-            start_color();
-            cbreak();
-            noecho();
-            keypad(stdscr, TRUE);
-            win = newwin(20, 50, 1, 2);
-            keypad(win, TRUE);
-            wclear(win);
-            wrefresh(win);
-            kls_showList_toWin(temporary_kls, win);
-            delwin(win);
-            endwin();
-
-            fclose(kls_file);
-        }
-        break;
-        case 'K': {
-            char path_to_kls_file[600];
-            char static_path[500];
-            // Set static_path value to the correct static dir path
-            resolve_staticPath(static_path);
-
-            //Append to "kls_log.txt"
-            sprintf(path_to_kls_file, "%s/%s", static_path,
-                    "debug_log.txt");
-            FILE *kls_file = NULL;
-            kls_file = fopen(path_to_kls_file, "a");
-            if (kls_file == NULL) {
-                fprintf(stderr,
-                        "debug_generic():  failed opening debug logfile.\n");
-                exit(EXIT_FAILURE);
-            }
-            if (kls == NULL) {
-                fprintf(stderr, "debug_generic():  kls was NULL.\n");
-                exit(EXIT_FAILURE);
-            }
-
-            fprintf(kls_file, "--BEGIN debug of passed kls--\n");
-            for (int i = HR_Path; i < HLP_MAX_INDEX + 99; i++) {
-                ptrdiff_t usage = kls_type_usage(i, kls);
-#ifndef _WIN32
-                fprintf(kls_file,
-                        "Usage for HLP_Region_Type { %s } [Index: %i]  {Size: %li }\n",
-                        stringFrom_HLP_Region_Type(i - 101 +
-                                                   KLS_REGIONTYPE_MAX), i,
-                        usage);
-#else
-                fprintf(kls_file,
-                        "Usage for HLP_Region_Type { %s } [Index: %i]  {Size: %lli }\n",
-                        stringFrom_HLP_Region_Type(i - 101 +
-                                                   KLS_REGIONTYPE_MAX), i,
-                        usage);
-#endif
-            }
-            fprintf(kls_file, "--END debug of passed kls--\n\n");
-
-            fclose(kls_file);
-        }
-        break;
-        case 'k': {
-            char path_to_kls_file[600];
-            char static_path[500];
-            // Set static_path value to the correct static dir path
-            resolve_staticPath(static_path);
-
-            //Append to "kls_log.txt"
-            sprintf(path_to_kls_file, "%s/%s", static_path,
-                    "debug_log.txt");
-            FILE *kls_file = NULL;
-            kls_file = fopen(path_to_kls_file, "a");
-            if (kls_file == NULL) {
-                fprintf(stderr,
-                        "debug_generic():  failed opening debug logfile.\n");
-                exit(EXIT_FAILURE);
-            }
-            if (kls == NULL) {
-                fprintf(stderr, "debug_generic():  kls was NULL.\n");
-                exit(EXIT_FAILURE);
-            }
-            fprintf(kls_file, "--BEGIN debug of passed kls--\n");
-            print_kls_2file(kls_file, kls);
-            kls_showList_toFile(kls->regs, kls_file);
-            kls_usageReport_toFile(kls, kls_file);
-            fprintf(kls_file, "--END debug of passed kls--\n\n");
-            WINDOW *win = NULL;
-            /* Initialize curses */
-            clear();
-            refresh();
-            start_color();
-            cbreak();
-            noecho();
-            keypad(stdscr, TRUE);
-            win = newwin(20, 50, 1, 2);
-            keypad(win, TRUE);
-            wclear(win);
-            wrefresh(win);
-            kls_showList_toWin(kls, win);
-            delwin(win);
-            endwin();
-
-            fclose(kls_file);
-        }
-        break;
-        case 'Q': {
-            G_FASTQUIT_ON = (G_FASTQUIT_ON == 1 ? 0 : 1);
-            sprintf(msg, "Toggled G_FASTQUIT_ON, new value: (%i)",
-                    G_FASTQUIT_ON);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 'g': {
-            G_GODMODE_ON = (G_GODMODE_ON == 1 ? 0 : 1);
-            sprintf(msg, "Toggled G_GODMODE_ON, new value: (%i)",
-                    G_GODMODE_ON);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 'L': {
-            G_LOG_ON = (G_LOG_ON == 1 ? 0 : 1);
-            sprintf(msg, "Toggled G_LOG_ON, new value: (%i)", G_LOG_ON);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        case 's': {		//Query sprites slideshow
-            picked_debug_proc = 1;
-            int res = system("clear");
-            sprintf(msg, "debug_generic() 4 system(\"clear\") res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            int comm_len = -1;
-            int picked_comm = 0;
-            char comm[15] = "empty";
-            char *desc = NULL;
-            int sprite_total = -1;
-            int tot_cons = CONSUMABLESMAX + 1;
-            int tot_enemy = ENEMYCLASSESMAX + 1;
-            int tot_boss = BOSSCLASSESMAX + 1;
-            int tot_art = ARTIFACTSMAX + 1;
-            int tot_eq = EQUIPSMAX + 1;
-            int tot_eqzones = EQUIPZONES + 1;
-            int tot_misc = MISC_SPRITES_MAX + 1;
-            int allsprites =
-                tot_cons +
-                tot_enemy +
-                tot_boss + tot_art + tot_eq + tot_eqzones + tot_misc;
-
-            printf("\nConsumables sprites: \t%d\
-\nArtifacts sprites: \t%d\
-\nEquips sprites: \t%d\
-\nEquip zone sprites: \t%d\
-\nEnemies sprites: \t%d\
-\nBosses sprites: \t%d\
-\nOther misc sprites: \t%d\n", tot_cons, tot_art, tot_eq, tot_eqzones, tot_enemy, tot_boss, tot_misc);
-            printf("--------------------------");
-            printf("\nTotal sprites: \t\t%d\n", allsprites);
-            printf
-            ("\nInsert sprites query, q to quit: (q|boss|cons|equip|eqzone|artf|enemy|misc):\n\n[%s@spriteshow]$ ",
-             player->name);
-            while (!picked_comm && (comm_len = scanf("%10s", comm)) > -1) {
-                int res = scanf("%*c");
-                sprintf(msg, "debug_generic() 7 scanf() res was (%i)", res);
-                log_tag("debug_log.txt", "[DEBUG]", msg);
-                if (strcmp(comm, "q") == 0) {
-                    return;
-                } else if (strcmp(comm, "boss") == 0) {
-                    picked_comm = 1;
-                    sprite_total = BOSSCLASSESMAX + 1;
-                } else if (strcmp(comm, "cons") == 0) {
-                    picked_comm = 1;
-                    sprite_total = CONSUMABLESMAX + 1;
-                } else if (strcmp(comm, "equip") == 0) {
-                    picked_comm = 1;
-                    sprite_total = EQUIPSMAX + 1;
-                } else if (strcmp(comm, "eqzone") == 0) {
-                    picked_comm = 1;
-                    sprite_total = EQUIPZONES + 1;
-                } else if (strcmp(comm, "artf") == 0) {
-                    picked_comm = 1;
-                    sprite_total = ARTIFACTSMAX + 1;
-                } else if (strcmp(comm, "enemy") == 0) {
-                    picked_comm = 1;
-                    sprite_total = ENEMYCLASSESMAX + 1;
-                } else if (strcmp(comm, "misc") == 0) {
-                    picked_comm = 1;
-                    sprite_total = MISC_SPRITES_MAX + 1;
-                } else {
-                    printf
-                    ("\nInsert sprites query, q to quit: (q|boss|cons|equip|eqzone|artf|enemy|misc):\n\n[%s@spriteshow]$ ",
-                     player->name);
-                }
-            }
-            int sprite_count = -1;
-            int c = -1;
-            char s[20];
-            int startx = 0;
-            int x, y;
-            WINDOW *spriteshow_win;
-            /* Initialize curses */
-            //initscr();
-            start_color();
-            clear();
-            refresh();
-            cbreak();
-            noecho();
-            keypad(stdscr, TRUE);
-            spriteshow_win = newwin(19, 19, 2, 2);
-            keypad(spriteshow_win, TRUE);
-            for (sprite_count = 0; sprite_count < sprite_total;
-                 sprite_count++) {
-                if (strcmp(comm, "boss") == 0) {
-                    desc = stringFromBossClass(sprite_count);
-                } else if (strcmp(comm, "cons") == 0) {
-                    desc = stringFromConsumables(sprite_count);
-                } else if (strcmp(comm, "equip") == 0) {
-                    desc = stringFromEquips(sprite_count);
-                } else if (strcmp(comm, "eqzone") == 0) {
-                    desc = stringFromEquipzones(sprite_count);
-                } else if (strcmp(comm, "artf") == 0) {
-                    desc = stringFromArtifacts(sprite_count);
-                } else if (strcmp(comm, "enemy") == 0) {
-                    desc = stringFromEClass(sprite_count);
-                } else if (strcmp(comm, "misc") == 0) {
-                    desc = "Misc";
-                }
-                wclear(spriteshow_win);
-                wrefresh(spriteshow_win);
-                wclear(stdscr);
-                wrefresh(stdscr);
-                sprintf(msg, "Showing sprite n.%d, class is: %s\n",
-                        sprite_count, desc);
-                log_tag("debug_log.txt", "[DEBUG]", msg);
-                mvwprintw(stdscr, 5, 20, "Sprite for: \'%s\'", desc);
-                mvwprintw(stdscr, 7, 20, "(%i/%i)", sprite_count + 1,
-                          sprite_total);
-                refresh();
-                for (int i = 0; i < 8; i++) {
-                    if (strcmp(comm, "boss") == 0) {
-                        strcpy(s, bosses_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "cons") == 0) {
-                        strcpy(s, consumables_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "equip") == 0) {
-                        strcpy(s, equips_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "eqzone") == 0) {
-                        strcpy(s, equipzones_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "artf") == 0) {
-                        strcpy(s, artifacts_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "enemy") == 0) {
-                        strcpy(s, enemies_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "misc") == 0) {
-                        strcpy(s, misc_sprites[sprite_count][i]);
-                    }
-                    //sprintf(msg,"[DEBUG]    Copied string: \'%s\'\n", s);
-                    //debug_log("debug_log.txt",msg);
-
-                    for (int j = 0; j < 13; j++) {
-                        x = startx + 1 + j;
-                        y = 1 + i;
-
-                        print_encoded_char(spriteshow_win, y, x, s[j]);
-                    }	//End for line print
-                }		//End for sprite print
-
-                wrefresh(spriteshow_win);
-                //printf("\nPress Enter for next, q to quit.\n");
-                int picked = 0;
-                int go_previous = 0;
-                int quit_show = 0;
-                while (!picked
-                       && ((c = wgetch(spriteshow_win)) != KEY_F(1))) {
-                    if (c == 10 || c == KEY_RIGHT) {	//Enter, right
-                        sprintf(msg, "Going to next sprite.\n");
-                        log_tag("debug_log.txt", "[DEBUG]", msg);
-                        wclear(spriteshow_win);
-                        wrefresh(spriteshow_win);
-                        picked = 1;
-                        continue;	//Go to next sprite
-                    } else if (c == 'q') {	//Quit
-                        sprintf(msg, "Stopping query sprite show.\n");
-                        log_tag("debug_log.txt", "[DEBUG]", msg);
-                        picked = 1;
-                        quit_show = 1;
-                    } else if (c == KEY_LEFT) {	//Go to previous sprite
-                        go_previous = 1;
-                        sprintf(msg, "Going to previous sprite.\n");
-                        log_tag("debug_log.txt", "[DEBUG]", msg);
-                        picked = 1;
-                    } else {	//Unexpected char
-                        sprintf(msg,
-                                "Wrong operation. Continuing with next sprite.\n");
-                        log_tag("debug_log.txt", "\033[1;31m[ERROR]", msg);
-                        picked = 1;
-                        continue;	//We still go to next sprite
-                    }
-                }		//End while wait for user input
-                if (go_previous) {
-                    sprite_count--;
-                    sprite_count--;
-                    if (sprite_count < -1)
-                        sprite_count = -1;
-                    go_previous = 0;
-                }
-                if (quit_show) {
-                    break;
-                }
-            }		//End for all sprites
-            sprintf(msg, "End of sprite show.\n");
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-
-            delwin(spriteshow_win);
-            endwin();
-        }
-        break;
-        case 'q': {
-            return;
-        }
-        default: {		//Echo the passed char and ask for one more.
-            char cmd[50];
-            sprintf(cmd, "\necho \"%c\\n\"\n%c\n\n", ch[0], ch[0]);
-            printf("%s", cmd);
-            napms(500);
-        }
-        break;
-        }			//Close switch on ch[0]
-    }				//Close while !picked_debug_proc
-
-    res = system("clear");
-    sprintf(msg, "debug_generic() final system(\"clear\") res was (%i)", res);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-}
-
-/**
- * Takes a Room, a Fighter, a Enemy and a Path pointers (and integers for current room and enemy indexes) and asks user input to execute debug actions.
- * @see Room
- * @see Fighter
- * @see Enemy
- * @see Path
- * @see checkremainder()
- * @see statReset()
- * @see statResetEnemy()
- * @see GET_CALLBACK()
- * @see unlock_special()
- * @see printCounters()
- * @see dropEquip()
- * @param gmst The Gamestate pointer.
- * @param player The Fighter pointer at hand.
- * @param e The Enemy pointer for current enemy.
- * @param p The Path pointer of the current game.
- * @param roomIndex The index of current room.
- * @param currentEnemyNum The index of current enemy.
- * @param kls The Koliseo used for allocations.
- * @param t_kls The Koliseo_Temp used for temporary allocations.
- */
-void debug_enemies_room(Gamestate *gmst, Room *room, Fighter *player, Enemy *e,
-                        Path *p, int roomIndex, int currentEnemyNum,
-                        Koliseo *kls, Koliseo_Temp *t_kls)
-{
-
-    char msg[200];
-    char ch[25];
-    int picked_debug_proc = 0;
-#ifndef _WIN32
-    struct utsname uts;
-    uname(&uts);
-    sprintf(msg, "debug_enemies_room() loaded utsname using uname().\n");
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-    sprintf(msg, "System is %s\n", uts.sysname);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-    sprintf(msg, "OS Release is %s\n", uts.release);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-    sprintf(msg, "OS Version is %s\n", uts.version);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-    sprintf(msg, "Machine is %s\n", uts.machine);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-#endif
-
-    int res = system("clear");
-    sprintf(msg, "debug_enemies_room() system(\"clear\") res was (%i)", res);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-
-    int c = 0, n = -1;
-    while (!picked_debug_proc) {
-        int res = system("clear");
-        sprintf(msg, "debug_enemies_room() 2 system(\"clear\") res was (%i)",
-                res);
-        log_tag("debug_log.txt", "[DEBUG]", msg);
-        printf("\n\
-	What do you want to do? ('q' to quit)\n\
-		[\n\
-			'0':  Give xp points\t'1':  Give consumable\n\
-			'2':  Reset stats\t'3':  Reset Enemy\n\
-			'4':  Alter luck\t'5':  Give coins\n\
-			'6':  Unlock special\t'7':  Unlock Artifact\n\
-			'8':  Print counters\t'9':  Give Equip\n\
-			's':  Sprites slideshow\t'd': Dump debug symbols\n\
-			'f':  Show foes info\t'g': Toggle godmode\n\
-			'A':  Toggle autosave\t'Q': Toggle fast quit\n\
-			'L':  Toggle logging\t'k': Log passed Koliseo info\n\
-			't': Log global temporary_kls Koliseo info\n\
-			'K': Log usage stats for passed kls to debug log file.\n\
-			'T': Log usage stats for temporary_kls to debug log file.\n\
-			'G': Log Gamestate to debug log file.\n\
-			'q': Quit\t{Return}  Process your input line.\n\
-		]\n\n\
-	[%s@debug-func]$ ", player->name);
-
-        char *fgetsres = fgets(ch, sizeof ch, stdin);
-        sprintf(msg, "debug_enemies_room() fgets() res was (%s)", fgetsres);
-        log_tag("debug_log.txt", "[DEBUG]", msg);
-        switch (ch[0]) {
-        case '0': {
-            picked_debug_proc = 1;
-            int xp;
-            do {
-                printf("\nHow much?\n");
-            } while ((c = scanf("%i", &xp)) != 1 || xp < 0);
-            int res = scanf("%*c");
-            sprintf(msg, "debug_enemies_room() scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            checkremainder(player, xp);
-        }
-        break;
-        case '1': {
-            picked_debug_proc = 1;
-            int q = -1;
-            do {
-                printf("\nInsert consumable number:\n");
-            } while ((c = scanf("%i", &n)) != 1 || n > CONSUMABLESMAX
-                     || n < 0);
-            int res = scanf("%*c");
-            sprintf(msg, "debug_enemies_room() scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            do {
-                printf("\nInsert quantity:\n");
-            } while ((c = scanf("%i", &q)) != 1 && q <= 0);
-            res = scanf("%*c");
-            sprintf(msg, "debug_enemies_room() 2 scanf() res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            Consumable *c = (Consumable *) player->consumablesBag[n];
-            c->qty += q;
-            player->stats->consumablesfound += q;
-        }
-        break;
-        case '2': {
-            picked_debug_proc = 1;
-            statReset(player, 1);
-        }
-        break;
-        case '3': {
-            picked_debug_proc = 1;
-            statResetEnemy(e, 1);
-        }
-        break;
-        case '4': {
-            picked_debug_proc = 1;
-            printf("\nCurrent luck: %i\tRL: %i\n", player->luck, p->luck);
-
-            do {
-                printf("\nInsert new luck:\n");
-            } while ((c = scanf("%i", &n)) != 1 && n > 0);
-            int res = scanf("%*c");
-            sprintf(msg, "debug_enemies_room() 3 scanf() res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            //FIXME:
-            //Copy-pasted the calc for player luck... might need a function
-            p->luck = n;
-
-            player->luck = (p->luck * MAXPLAYERLUCK) / MAXLUCK;
-        }
-        break;
-        case '5': {
-            picked_debug_proc = 1;
-            /* Old code invoking the macro for special moves directly. Used for testing.
-             * int s = -1;
-             * printf("Insert special num:\n");
-             * scanf("%i",&s);
-             * scanf("%*c");
-             * printf("Read: %i\n",s);
-             * GET_CALLBACK(s,callback_special_t)(player,e,p,roomIndex,currentEnemyNum);
-             */
-
-            //printActivePerks(player);
-            int c;
-            int n = -1;
-            do {
-                printf("\nInsert coin number (0<n<100):\n");
-            } while ((c = scanf("%i", &n)) < 0 || c > 3 || n < 0
-                     || n > 100);
-            int res = scanf("%*c");
-            sprintf(msg, "debug_enemies_room() 4 scanf() res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            player->balance += n;
-        }
-        break;
-        case '6': {
-            picked_debug_proc = 1;
-            unlockSpecial(player);
-        }
-        break;
-        case '7': {
-            picked_debug_proc = 1;
-            int n = -1;
-            do {
-                printf("\nInsert artifact number (0<n<%i):\n",
-                       ARTIFACTSMAX + 1);
-            } while ((c = scanf("%i", &n)) != 1 || n > ARTIFACTSMAX
-                     || n < 0);
-
-            int res = scanf("%*c");
-            sprintf(msg, "debug_enemies_room() scanf() res was (%i)", res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            player->artifactsBag[n]->qty += 1;
-            player->artifactsBag[n]->active = 0;
-            player->stats->artifactsfound += 1;
-        }
-        break;
-        case '8': {
-            picked_debug_proc = 1;
-            int res = system("clear");
-            sprintf(msg,
-                    "debug_enemies_room() 3 system(\"clear\") res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            printf("\nPlayer Counters:\n");
-            printCounters((Turncounter **) player->counters);
-            printf("\nEnemy (%s) Counters:\n", stringFromEClass(e->class));
-            printCounters((Turncounter **) e->counters);
-            printf("\nPress Enter to resume game");
-            res = scanf("%*c");
-            sprintf(msg, "debug_enemies_room() 5 scanf() res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case '9': {
-            picked_debug_proc = 1;
-            int q = -1;
-            int c = -1;
-            do {
-                printf("\nInsert quantity:\n");
-            } while ((c = scanf("%i", &q)) != 1 && q <= 0);
-            int res = scanf("%*c");
-            sprintf(msg, "debug_enemies_room() 6 scanf() res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-
-            //TODO: handle dropping an equip with a new notification window
-            //for (int i=0; i<q;i++) {dropEquip(player,1,notify_win);};
-            log_tag("debug_log.txt", "[DEBUG]", "TODO: not working ATM.");
-        }
-        break;
-        case 'd': {
-            picked_debug_proc = 1;
-            printf("\nVERSION:    %s\n", VERSION);
-#ifndef _WIN32
-            printf("\nSystem:    %s\n", uts.sysname);
-            printf("\nOS Release:    %s\n", uts.release);
-            printf("\nOS Version:    %s\n", uts.version);
-            printf("\nMachine:    %s\n", uts.machine);
-#endif
-            printf("\nGAMEMODE:    %s\n", stringFromGamemode(GAMEMODE));
-            printf("\nPath->current_saveslot->save_path:    %s\n",
-                   p->current_saveslot->save_path);
-            printf("\nGS_AUTOSAVE_ON:    %i\n", GS_AUTOSAVE_ON);
-            printf("\nG_FASTQUIT_ON:    %i\n", G_FASTQUIT_ON);
-            printf("\nG_DEBUG_ON:    %i\n", G_DEBUG_ON);
-            printf("\nG_LOG_ON:    %i\n", G_LOG_ON);
-            printf("\nG_GODMODE_ON:    %i\n", G_GODMODE_ON);
-            printf("\nG_EXPERIMENTAL_ON:    %i\n", G_EXPERIMENTAL_ON);
-            printf("\nG_DEBUG_ROOMTYPE_ON:    %i\n", G_DEBUG_ROOMTYPE_ON);
-            if (G_DEBUG_ROOMTYPE_ON == 1) {
-                printf("\nG_DEBUG_ROOMTYPE:    %i\n", G_DEBUG_ROOMTYPE);
-            }
-            printf("\nG_DEBUG_ENEMYTYPE_ON:    %i\n", G_DEBUG_ENEMYTYPE_ON);
-            if (G_DEBUG_ENEMYTYPE_ON == 1) {
-                printf("\nG_DEBUG_ENEMYTYPE:    %i\n", G_DEBUG_ENEMYTYPE);
-            }
-            printf("\nPress Enter to resume game.\n");
-            int res = scanf("%*c");
-            sprintf(msg, "debug_enemies_room() 7 scanf() res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 'f': {
-            clear();
-            refresh();
-            debug_printFoeParty(room->foes);
-        }
-        break;
-        case 'G': {
-            log_tag("debug_log.txt", "[DEBUG]", "Logging Gamestate");
-            dbg_Gamestate(gmst);
-            log_tag("debug_log.txt", "[DEBUG]", "Done logging Gamestate");
-        }
-        break;
-        case 'T': {
-            char path_to_kls_file[600];
-            char static_path[500];
-            // Set static_path value to the correct static dir path
-            resolve_staticPath(static_path);
-
-            //Append to "kls_log.txt"
-            sprintf(path_to_kls_file, "%s/%s", static_path,
-                    "debug_log.txt");
-            FILE *kls_file = NULL;
-            kls_file = fopen(path_to_kls_file, "a");
-            if (kls_file == NULL) {
-                fprintf(stderr,
-                        "debug_generic():  failed opening debug logfile.\n");
-                exit(EXIT_FAILURE);
-            }
-            if (kls == NULL) {
-                fprintf(stderr, "debug_generic():  kls was NULL.\n");
-                exit(EXIT_FAILURE);
-            }
-
-            fprintf(kls_file, "--BEGIN debug of temporary_kls--\n");
-            for (int i = HR_Path; i < HLP_MAX_INDEX + 99; i++) {
-                ptrdiff_t usage = kls_type_usage(i, temporary_kls);
-#ifndef _WIN32
-                fprintf(kls_file,
-                        "Usage for HLP_Region_Type { %s } [Index: %i]  {Size: %li }\n",
-                        stringFrom_HLP_Region_Type(i - 101 +
-                                                   KLS_REGIONTYPE_MAX), i,
-                        usage);
-#else
-                fprintf(kls_file,
-                        "Usage for HLP_Region_Type { %s } [Index: %i]  {Size: %lli }\n",
-                        stringFrom_HLP_Region_Type(i - 101 +
-                                                   KLS_REGIONTYPE_MAX), i,
-                        usage);
-#endif
-            }
-            fprintf(kls_file, "--END debug of temporary_kls--\n\n");
-
-            fclose(kls_file);
-        }
-        break;
-        case 't': {
-            char path_to_kls_file[600];
-            char static_path[500];
-            // Set static_path value to the correct static dir path
-            resolve_staticPath(static_path);
-
-            //Append to "kls_log.txt"
-            sprintf(path_to_kls_file, "%s/%s", static_path,
-                    "debug_log.txt");
-            FILE *kls_file = NULL;
-            kls_file = fopen(path_to_kls_file, "a");
-            if (kls_file == NULL) {
-                fprintf(stderr,
-                        "debug_generic():  failed opening debug logfile.\n");
-                exit(EXIT_FAILURE);
-            }
-            if (kls == NULL) {
-                fprintf(stderr, "debug_generic():  kls was NULL.\n");
-                exit(EXIT_FAILURE);
-            }
-            fprintf(kls_file, "--BEGIN debug of temporary_kls--\n");
-            print_kls_2file(kls_file, temporary_kls);
-            kls_showList_toFile(temporary_kls->regs, kls_file);
-            kls_usageReport_toFile(temporary_kls, kls_file);
-            fprintf(kls_file, "--END debug of temporary_kls--\n\n");
-            WINDOW *win = NULL;
-            /* Initialize curses */
-            clear();
-            refresh();
-            start_color();
-            cbreak();
-            noecho();
-            keypad(stdscr, TRUE);
-            win = newwin(20, 50, 1, 2);
-            keypad(win, TRUE);
-            wclear(win);
-            wrefresh(win);
-            kls_showList_toWin(temporary_kls, win);
-            delwin(win);
-            endwin();
-
-            fclose(kls_file);
-        }
-        break;
-        case 'K': {
-            char path_to_kls_file[600];
-            char static_path[500];
-            // Set static_path value to the correct static dir path
-            resolve_staticPath(static_path);
-
-            //Append to "kls_log.txt"
-            sprintf(path_to_kls_file, "%s/%s", static_path,
-                    "debug_log.txt");
-            FILE *kls_file = NULL;
-            kls_file = fopen(path_to_kls_file, "a");
-            if (kls_file == NULL) {
-                fprintf(stderr,
-                        "debug_generic():  failed opening debug logfile.\n");
-                exit(EXIT_FAILURE);
-            }
-            if (kls == NULL) {
-                fprintf(stderr, "debug_generic():  kls was NULL.\n");
-                exit(EXIT_FAILURE);
-            }
-
-            fprintf(kls_file, "--BEGIN debug of passed kls--\n");
-            for (int i = HR_Path; i < HLP_MAX_INDEX + 99; i++) {
-                ptrdiff_t usage = kls_type_usage(i, kls);
-#ifndef _WIN32
-                fprintf(kls_file,
-                        "Usage for HLP_Region_Type { %s } [Index: %i]  {Size: %li }\n",
-                        stringFrom_HLP_Region_Type(i - 101 +
-                                                   KLS_REGIONTYPE_MAX), i,
-                        usage);
-#else
-                fprintf(kls_file,
-                        "Usage for HLP_Region_Type { %s } [Index: %i]  {Size: %lli }\n",
-                        stringFrom_HLP_Region_Type(i - 101 +
-                                                   KLS_REGIONTYPE_MAX), i,
-                        usage);
-#endif
-            }
-            fprintf(kls_file, "--END debug of passed kls--\n\n");
-
-            fclose(kls_file);
-        }
-        break;
-        case 'k': {
-            char path_to_kls_file[600];
-            char static_path[500];
-            // Set static_path value to the correct static dir path
-            resolve_staticPath(static_path);
-
-            //Append to "kls_log.txt"
-            sprintf(path_to_kls_file, "%s/%s", static_path,
-                    "debug_log.txt");
-            FILE *kls_file = NULL;
-            kls_file = fopen(path_to_kls_file, "a");
-            if (kls_file == NULL) {
-                fprintf(stderr,
-                        "debug_generic():  failed opening debug logfile.\n");
-                exit(EXIT_FAILURE);
-            }
-            if (kls == NULL) {
-                fprintf(stderr, "debug_generic():  kls was NULL.\n");
-                exit(EXIT_FAILURE);
-            }
-            fprintf(kls_file, "--BEGIN debug of passed kls--\n");
-            print_kls_2file(kls_file, kls);
-            kls_showList_toFile(kls->regs, kls_file);
-            kls_usageReport_toFile(kls, kls_file);
-            fprintf(kls_file, "--END debug of passed kls--\n\n");
-            WINDOW *win = NULL;
-            /* Initialize curses */
-            clear();
-            refresh();
-            start_color();
-            cbreak();
-            noecho();
-            keypad(stdscr, TRUE);
-            win = newwin(20, 50, 1, 2);
-            keypad(win, TRUE);
-            wclear(win);
-            wrefresh(win);
-            kls_showList_toWin(kls, win);
-            delwin(win);
-            endwin();
-            fclose(kls_file);
-        }
-        break;
-        case 'A': {
-            GS_AUTOSAVE_ON = (GS_AUTOSAVE_ON == 1 ? 0 : 1);
-            sprintf(msg, "Toggled G_AUTOSAVE_ON, new value: (%i)",
-                    GS_AUTOSAVE_ON);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 'Q': {
-            G_FASTQUIT_ON = (G_FASTQUIT_ON == 1 ? 0 : 1);
-            sprintf(msg, "Toggled G_FASTQUIT_ON, new value: (%i)",
-                    G_FASTQUIT_ON);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 'L': {
-            G_LOG_ON = (G_LOG_ON == 1 ? 0 : 1);
-            sprintf(msg, "Toggled G_LOG_ON, new value: (%i)", G_LOG_ON);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 'g': {
-            G_GODMODE_ON = (G_GODMODE_ON == 1 ? 0 : 1);
-            sprintf(msg, "Toggled G_GODMODE_ON, new value: (%i)",
-                    G_GODMODE_ON);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-        }
-        break;
-        case 's': {		//Query sprites slideshow
-            picked_debug_proc = 1;
-            int res = system("clear");
-            sprintf(msg,
-                    "debug_enemies_room() 4 system(\"clear\") res was (%i)",
-                    res);
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-            int comm_len = -1;
-            int picked_comm = 0;
-            char comm[15] = "empty";
-            char *desc = NULL;
-            int sprite_total = -1;
-            int tot_cons = CONSUMABLESMAX + 1;
-            int tot_enemy = ENEMYCLASSESMAX + 1;
-            int tot_boss = BOSSCLASSESMAX + 1;
-            int tot_art = ARTIFACTSMAX + 1;
-            int tot_eq = EQUIPSMAX + 1;
-            int tot_eqzones = EQUIPZONES + 1;
-            int tot_misc = MISC_SPRITES_MAX + 1;
-            int allsprites =
-                tot_cons +
-                tot_enemy +
-                tot_boss + tot_art + tot_eq + tot_eqzones + tot_misc;
-
-            printf("\nConsumables sprites: \t%d\
-\nArtifacts sprites: \t%d\
-\nEquips sprites: \t%d\
-\nEquip zone sprites: \t%d\
-\nEnemies sprites: \t%d\
-\nBosses sprites: \t%d\
-\nOther misc sprites: \t%d\n", tot_cons, tot_art, tot_eq, tot_eqzones, tot_enemy, tot_boss, tot_misc);
-            printf("--------------------------");
-            printf("\nTotal sprites: \t\t%d\n", allsprites);
-            printf
-            ("\nInsert sprites query, q to quit: (q|boss|cons|equip|eqzone|artf|enemy|misc):\n\n[%s@spriteshow]$ ",
-             player->name);
-            while (!picked_comm && (comm_len = scanf("%10s", comm)) > -1) {
-                int res = scanf("%*c");
-                sprintf(msg, "debug_enemies_room() 7 scanf() res was (%i)",
-                        res);
-                log_tag("debug_log.txt", "[DEBUG]", msg);
-                if (strcmp(comm, "q") == 0) {
-                    return;
-                } else if (strcmp(comm, "boss") == 0) {
-                    picked_comm = 1;
-                    sprite_total = BOSSCLASSESMAX + 1;
-                } else if (strcmp(comm, "cons") == 0) {
-                    picked_comm = 1;
-                    sprite_total = CONSUMABLESMAX + 1;
-                } else if (strcmp(comm, "equip") == 0) {
-                    picked_comm = 1;
-                    sprite_total = EQUIPSMAX + 1;
-                } else if (strcmp(comm, "eqzone") == 0) {
-                    picked_comm = 1;
-                    sprite_total = EQUIPZONES + 1;
-                } else if (strcmp(comm, "artf") == 0) {
-                    picked_comm = 1;
-                    sprite_total = ARTIFACTSMAX + 1;
-                } else if (strcmp(comm, "enemy") == 0) {
-                    picked_comm = 1;
-                    sprite_total = ENEMYCLASSESMAX + 1;
-                } else if (strcmp(comm, "misc") == 0) {
-                    picked_comm = 1;
-                    sprite_total = MISC_SPRITES_MAX + 1;
-                } else {
-                    printf
-                    ("\nInsert sprites query, q to quit: (q|boss|cons|equip|eqzone|artf|enemy|misc):\n\n[%s@spriteshow]$ ",
-                     player->name);
-                }
-            }
-            int sprite_count = -1;
-            int c = -1;
-            char s[20];
-            int startx = 0;
-            int x, y;
-            WINDOW *spriteshow_win;
-            /* Initialize curses */
-            //initscr();
-            start_color();
-            clear();
-            refresh();
-            cbreak();
-            noecho();
-            keypad(stdscr, TRUE);
-            spriteshow_win = newwin(19, 19, 2, 2);
-            keypad(spriteshow_win, TRUE);
-            for (sprite_count = 0; sprite_count < sprite_total;
-                 sprite_count++) {
-                if (strcmp(comm, "boss") == 0) {
-                    desc = stringFromBossClass(sprite_count);
-                } else if (strcmp(comm, "cons") == 0) {
-                    desc = stringFromConsumables(sprite_count);
-                } else if (strcmp(comm, "equip") == 0) {
-                    desc = stringFromEquips(sprite_count);
-                } else if (strcmp(comm, "eqzone") == 0) {
-                    desc = stringFromEquipzones(sprite_count);
-                } else if (strcmp(comm, "artf") == 0) {
-                    desc = stringFromArtifacts(sprite_count);
-                } else if (strcmp(comm, "enemy") == 0) {
-                    desc = stringFromEClass(sprite_count);
-                } else if (strcmp(comm, "misc") == 0) {
-                    desc = "Misc";
-                }
-                wclear(spriteshow_win);
-                wrefresh(spriteshow_win);
-                wclear(stdscr);
-                wrefresh(stdscr);
-                sprintf(msg, "Showing sprite n.%d, class is: %s\n",
-                        sprite_count, desc);
-                log_tag("debug_log.txt", "[DEBUG]", msg);
-                mvwprintw(stdscr, 5, 20, "Sprite for: \'%s\'", desc);
-                mvwprintw(stdscr, 7, 20, "(%i/%i)", sprite_count + 1,
-                          sprite_total);
-                refresh();
-                for (int i = 0; i < 8; i++) {
-                    if (strcmp(comm, "boss") == 0) {
-                        strcpy(s, bosses_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "cons") == 0) {
-                        strcpy(s, consumables_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "equip") == 0) {
-                        strcpy(s, equips_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "eqzone") == 0) {
-                        strcpy(s, equipzones_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "artf") == 0) {
-                        strcpy(s, artifacts_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "enemy") == 0) {
-                        strcpy(s, enemies_sprites[sprite_count][i]);
-                    } else if (strcmp(comm, "misc") == 0) {
-                        strcpy(s, misc_sprites[sprite_count][i]);
-                    }
-                    //sprintf(msg,"[DEBUG]    Copied string: \'%s\'\n", s);
-                    //debug_log("debug_log.txt",msg);
-
-                    for (int j = 0; j < 13; j++) {
-                        x = startx + 1 + j;
-                        y = 1 + i;
-
-                        print_encoded_char(spriteshow_win, y, x, s[j]);
-                    }	//End for line print
-                }		//End for sprite print
-
-                wrefresh(spriteshow_win);
-                //printf("\nPress Enter for next, q to quit.\n");
-                int picked = 0;
-                int go_previous = 0;
-                int quit_show = 0;
-                while (!picked
-                       && ((c = wgetch(spriteshow_win)) != KEY_F(1))) {
-                    if (c == 10 || c == KEY_RIGHT) {	//Enter, right
-                        sprintf(msg, "Going to next sprite.\n");
-                        log_tag("debug_log.txt", "[DEBUG]", msg);
-                        wclear(spriteshow_win);
-                        wrefresh(spriteshow_win);
-                        picked = 1;
-                        continue;	//Go to next sprite
-                    } else if (c == 'q') {	//Quit
-                        sprintf(msg, "Stopping query sprite show.\n");
-                        log_tag("debug_log.txt", "[DEBUG]", msg);
-                        picked = 1;
-                        quit_show = 1;
-                    } else if (c == KEY_LEFT) {	//Go to previous sprite
-                        go_previous = 1;
-                        sprintf(msg, "Going to previous sprite.\n");
-                        log_tag("debug_log.txt", "[DEBUG]", msg);
-                        picked = 1;
-                    } else {	//Unexpected char
-                        sprintf(msg,
-                                "Wrong operation. Continuing with next sprite.\n");
-                        log_tag("debug_log.txt", "\033[1;31m[ERROR]", msg);
-                        picked = 1;
-                        continue;	//We still go to next sprite
-                    }
-                }		//End while wait for user input
-                if (go_previous) {
-                    sprite_count--;
-                    sprite_count--;
-                    if (sprite_count < -1)
-                        sprite_count = -1;
-                    go_previous = 0;
-                }
-                if (quit_show) {
-                    break;
-                }
-            }		//End for all sprites
-            sprintf(msg, "End of sprite show.\n");
-            log_tag("debug_log.txt", "[DEBUG]", msg);
-
-            delwin(spriteshow_win);
-            endwin();
-        }
-        break;
-        case 'q': {
-            return;
-        }
-        default: {		//Echo the passed char and ask for one more.
-            char cmd[50];
-            sprintf(cmd, "\necho \"%c\\n\"\n%c\n\n", ch[0], ch[0]);
-            printf("%s", cmd);
-            napms(500);
-        }
-        break;
-        }			//Close switch on ch[0]
-    }				//Close while !picked_debug_proc
-
-}
-
-/**
- * Takes a Fighter, a Room and a loadInfo pointers, and prints fighter stats and a quitting message, before quitting the program and freeing Room.
- *
- * @see Fighter
- * @see printStats()
- * @see death()
- * @param p The Fighter pointer at hand.
- * @param room The Room pointer at hand.
- * @param load_info The loadInfo pointer at hand.
- * @param t_kls The Koliseo_Temp to end if possible.
- */
-void quit(Fighter *p, Room *room, loadInfo *load_info, Koliseo_Temp *t_kls)
-{
-    char msg[500];
-    Koliseo *kls = t_kls->kls;
-    endwin();
-    int res = system("reset");
-    sprintf(msg, "quit() system(\"reset\") res was (%i)", res);
-    log_tag("debug_log.txt", "[DEBUG]", msg);
-    //printf("\n\n\tTHANKS 4 PLAYING\n");
-    //FIXME
-    //dropping out of the Koliseo scope might render stat pointer invalid.
-    //Can't we print stats and clear the kls?
-    //printStats(p);
-    //printf("\n");
-#ifndef _WIN32
-    sprintf(msg, "Resetting Koliseo_Temp from: (%li)", t_kls->kls->offset);
-#else
-    sprintf(msg, "Resetting Koliseo_Temp from: (%lli)", t_kls->kls->offset);
-#endif
-    kls_log(kls, "DEBUG", msg);
-    death(p, load_info);
-    //FIXME
-    //I think we should free those?
-    //May cause a crash?
-    //kls_free(default_kls);
-    //kls_free(temporary_kls);
-    //FIXME:
-    //Calling this segfaults?
-    //freeRoom(room);
-    log_tag("debug_log.txt", "[DEBUG]", "Quitting program.");
-    exit(EXIT_SUCCESS);
 }
 
 /**
