+++ src/rooms.h
@@ -18,9 +18,8 @@
 
 #ifndef GAME_ROOMS_H
 #define GAME_ROOMS_H
-#include "helapordo.h"
-//#include "game_init.h"
-//#include "artifacts.h"
+#include "artifacts.h"
+#include "turn_op.h"
 
 void initRoom(Room * r, Fighter * f, int index, roomClass type, int enemyTotal,
               loadInfo * load_info, Koliseo_Temp * t_kls);
